Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult

A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n = 100, m = 8, betaSD = 1)
dds <- DESeq2::DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- DESeq2::results(dds)
deseqresult2tbl(res)
#>              id     baseMean log2FoldChange     lfcSE        stat       pvalue
#> gene2     gene2  889.0763443     1.40457360 0.2851405  4.92589965 8.397309e-07
#> gene10   gene10   31.2939691    -2.30781587 0.4966262 -4.64698775 3.368172e-06
#> gene34   gene34  292.4783767     1.64147411 0.3997003  4.10676205 4.012442e-05
#> gene41   gene41   28.6901101     2.21845290 0.5836904  3.80073556 1.442672e-04
#> gene88   gene88   60.3159852     1.37716689 0.3659257  3.76351542 1.675414e-04
#> gene56   gene56    8.3604780    -3.27936050 1.0035856 -3.26764402 1.084467e-03
#> gene68   gene68   27.7497242     2.05751127 0.6709718  3.06646439 2.166066e-03
#> gene29   gene29   19.5178067    -1.67837782 0.5554410 -3.02170314 2.513569e-03
#> gene75   gene75   42.7995221    -1.41309046 0.4807749 -2.93919375 3.290673e-03
#> gene36   gene36    6.8045022    -2.88289840 1.0059953 -2.86571765 4.160652e-03
#> gene72   gene72   17.4561303    -1.82840989 0.6337666 -2.88498925 3.914270e-03
#> gene11   gene11   47.2227857     1.44647249 0.5460647  2.64890317 8.075347e-03
#> gene100 gene100   64.0888817     1.09601072 0.4280672  2.56037083 1.045605e-02
#> gene95   gene95 1846.6668183    -0.61881493 0.2497159 -2.47807630 1.320929e-02
#> gene46   gene46  667.6238370     0.85461677 0.3536107  2.41682934 1.565636e-02
#> gene40   gene40   32.4741975    -1.26895196 0.5306946 -2.39111525 1.679728e-02
#> gene1     gene1   38.2643998    -1.24354093 0.5670404 -2.19303777 2.830466e-02
#> gene9     gene9  416.6756303     0.80231002 0.3650207  2.19798513 2.795017e-02
#> gene24   gene24    5.6926030     2.99397364 1.3179470  2.27169507 2.310493e-02
#> gene37   gene37  155.2345368    -0.78120106 0.3503636 -2.22968687 2.576824e-02
#> gene76   gene76   31.9591151     1.01489927 0.4644571  2.18513018 2.887930e-02
#> gene80   gene80  105.4338473     0.87256619 0.3911515  2.23076255 2.569686e-02
#> gene92   gene92   10.9090922     1.88425258 0.8406294  2.24147836 2.499511e-02
#> gene49   gene49  210.6249199    -0.68632162 0.3537769 -1.93998444 5.238158e-02
#> gene65   gene65    7.1093476    -1.57457453 0.8710469 -1.80768055 7.065623e-02
#> gene33   gene33   12.3374822     1.55494426 0.8764584  1.77412206 7.604296e-02
#> gene63   gene63   23.1290067     0.95257781 0.5481516  1.73779989 8.224608e-02
#> gene20   gene20    9.4416374    -1.34141327 0.8002775 -1.67618524 9.370191e-02
#> gene23   gene23   15.0678979    -0.95897088 0.5693045 -1.68446037 9.209273e-02
#> gene93   gene93  182.6820001     0.49075360 0.3046673  1.61078539 1.072265e-01
#> gene48   gene48   22.5569637    -0.79869262 0.5255478 -1.51973354 1.285780e-01
#> gene58   gene58   29.0107265    -0.78269410 0.5110534 -1.53153098 1.256382e-01
#> gene89   gene89  148.4380736     0.50009844 0.3425828  1.45978872 1.443481e-01
#> gene52   gene52    9.6067344     1.08727383 0.7678056  1.41607957 1.567522e-01
#> gene66   gene66   25.2733674     0.83020838 0.5938230  1.39807389 1.620909e-01
#> gene5     gene5   18.8803870     0.70879844 0.5809023  1.22016800 2.224012e-01
#> gene78   gene78   18.8104041     1.70515538 1.4314867  1.19117793 2.335837e-01
#> gene14   gene14   13.2592507    -0.82445111 0.7294649 -1.13021362 2.583862e-01
#> gene39   gene39   20.5579306    -0.67538739 0.5962972 -1.13263559 2.573673e-01
#> gene64   gene64   13.6007577    -0.69577811 0.6080430 -1.14429100 2.525030e-01
#> gene51   gene51    7.9153176     0.89053503 0.8097773  1.09972829 2.714505e-01
#> gene4     gene4   31.6880524    -0.62726918 0.5856031 -1.07115076 2.841016e-01
#> gene30   gene30  109.4828789    -0.34029759 0.3315114 -1.02650332 3.046544e-01
#> gene97   gene97   13.3462163     0.70794810 0.7017257  1.00886731 3.130383e-01
#> gene13   gene13   39.9005392    -0.38671503 0.4133575 -0.93554625 3.495069e-01
#> gene15   gene15  195.6470114     0.29383404 0.3116309  0.94289132 3.457365e-01
#> gene71   gene71   56.8913600    -0.32800941 0.3505925 -0.93558591 3.494864e-01
#> gene16   gene16   10.5995669     0.70299583 0.7825916  0.89829203 3.690299e-01
#> gene28   gene28   11.1572969    -0.77892910 0.8696706 -0.89565993 3.704344e-01
#> gene57   gene57   91.3595208     0.34878880 0.3938402  0.88560990 3.758278e-01
#> gene32   gene32   18.3722380    -0.59799557 0.7007241 -0.85339659 3.934394e-01
#> gene96   gene96   13.6880873     0.52698379 0.6495230  0.81133976 4.171706e-01
#> gene60   gene60    8.0585681    -0.53501821 0.7519197 -0.71153638 4.767519e-01
#> gene81   gene81   16.2356039    -0.45089179 0.6712091 -0.67176056 5.017361e-01
#> gene99   gene99   63.4198441     0.28908707 0.4314691  0.67000641 5.028537e-01
#> gene69   gene69    8.5104551     0.57648757 0.9173064  0.62845692 5.297047e-01
#> gene77   gene77  184.5030088     0.23674653 0.3965570  0.59700502 5.505040e-01
#> gene43   gene43   43.8416003    -0.23388278 0.4165551 -0.56146909 5.744778e-01
#> gene54   gene54   12.0782421    -0.36422930 0.6603840 -0.55154168 5.812624e-01
#> gene12   gene12   41.2874938     0.21761107 0.4222650  0.51534238 6.063138e-01
#> gene27   gene27   57.1916537    -0.24655989 0.4781400 -0.51566464 6.060887e-01
#> gene25   gene25   22.0150945     0.27169882 0.5780475  0.47002851 6.383346e-01
#> gene31   gene31    9.2448574     0.36552138 0.7845796  0.46588183 6.413001e-01
#> gene62   gene62   39.4512334     0.21223203 0.4940868  0.42954401 6.675274e-01
#> gene98   gene98    6.0195519    -0.39465405 0.9069293 -0.43515414 6.634505e-01
#> gene79   gene79   19.3932149     0.25444911 0.6785382  0.37499602 7.076634e-01
#> gene8     gene8  121.1553094    -0.11020925 0.3218386 -0.34243637 7.320225e-01
#> gene84   gene84   28.0375387    -0.15931934 0.5730881 -0.27800150 7.810112e-01
#> gene91   gene91   24.7384256     0.17784000 0.6295677  0.28247953 7.775758e-01
#> gene74   gene74   19.7003126    -0.12018379 0.5170926 -0.23242218 8.162101e-01
#> gene82   gene82   95.1746176     0.09687399 0.4030363  0.24036046 8.100508e-01
#> gene85   gene85    8.6613423    -0.12959318 0.7510858 -0.17254112 8.630121e-01
#> gene18   gene18   43.1689508    -0.05802110 0.4427113 -0.13105855 8.957290e-01
#> gene50   gene50   36.3614482    -0.06732547 0.4668478 -0.14421288 8.853324e-01
#> gene21   gene21   71.7561524     0.04966879 0.4317812  0.11503229 9.084195e-01
#> gene61   gene61    7.0728805    -0.06293746 0.9006009 -0.06988385 9.442861e-01
#> gene67   gene67   34.0649314     0.03613679 0.5006701  0.07217685 9.424612e-01
#> gene94   gene94    6.5632256     0.05451695 1.0055319  0.05421702 9.567623e-01
#> gene3     gene3    5.0783847    -0.52590855 1.0368722 -0.50720674 6.120098e-01
#> gene6     gene6    1.4160200    -4.08391942 2.0879744 -1.95592410 5.047410e-02
#> gene7     gene7    4.2624642    -0.28964227 1.1551606 -0.25073766 8.020169e-01
#> gene17   gene17    0.1139798     0.56436799 3.5338115  0.15970517 8.731133e-01
#> gene19   gene19    3.2007891    -1.84353615 1.2665876 -1.45551414 1.455270e-01
#> gene22   gene22    2.6978821     0.02430144 1.5131176  0.01606051 9.871861e-01
#> gene26   gene26    0.0000000             NA        NA          NA           NA
#> gene35   gene35    2.5177516     0.80061513 1.5012687  0.53329236 5.938312e-01
#> gene38   gene38    1.8513665    -0.75640627 1.4766505 -0.51224461 6.084798e-01
#> gene42   gene42    5.3001545    -0.49421708 1.0666836 -0.46332116 6.431342e-01
#> gene44   gene44    1.3761684     3.04049068 2.0149890  1.50893664 1.313150e-01
#> gene45   gene45    3.7021365    -0.38667401 1.4762237 -0.26193456 7.933719e-01
#> gene47   gene47    1.1533728    -2.04058398 2.2295749 -0.91523455 3.600685e-01
#> gene53   gene53    0.9147058    -0.72142534 2.7269811 -0.26455091 7.913554e-01
#> gene55   gene55    3.5677243     0.09832137 1.1769142  0.08354167 9.334208e-01
#> gene59   gene59    5.4184385     0.61062645 1.1889282  0.51359406 6.075358e-01
#> gene70   gene70    5.3736856    -1.60068602 0.9813727 -1.63106834 1.028759e-01
#> gene73   gene73    5.2150589     0.71060797 1.0808416  0.65745801 5.108865e-01
#> gene83   gene83    3.5662663    -3.87528713 1.4354423 -2.69971633 6.939862e-03
#> gene86   gene86    5.1534628    -1.37305084 1.0176569 -1.34922762 1.772639e-01
#> gene87   gene87    2.4727188    -2.16978868 1.5963821 -1.35919132 1.740860e-01
#> gene90   gene90    0.6085449     0.48284029 2.9198850  0.16536278 8.686585e-01
#>                 padj
#> gene2   6.549901e-05
#> gene10  1.313587e-04
#> gene34  1.043235e-03
#> gene41  2.613646e-03
#> gene88  2.613646e-03
#> gene56  1.409807e-02
#> gene68  2.413616e-02
#> gene29  2.450730e-02
#> gene75  2.851917e-02
#> gene36  2.950280e-02
#> gene72  2.950280e-02
#> gene11  5.248975e-02
#> gene100 6.273631e-02
#> gene95  7.359461e-02
#> gene46  8.141305e-02
#> gene40  8.188674e-02
#> gene1   9.793850e-02
#> gene9   9.793850e-02
#> gene24  9.793850e-02
#> gene37  9.793850e-02
#> gene76  9.793850e-02
#> gene80  9.793850e-02
#> gene92  9.793850e-02
#> gene49  1.702401e-01
#> gene65  2.204475e-01
#> gene33  2.281289e-01
#> gene63  2.375998e-01
#> gene20  2.520258e-01
#> gene23  2.520258e-01
#> gene93  2.787889e-01
#> gene48  3.134088e-01
#> gene58  3.134088e-01
#> gene89  3.411865e-01
#> gene52  3.596080e-01
#> gene66  3.612311e-01
#> gene5   4.818693e-01
#> gene78  4.924198e-01
#> gene14  5.038531e-01
#> gene39  5.038531e-01
#> gene64  5.038531e-01
#> gene51  5.164181e-01
#> gene4   5.276173e-01
#> gene30  5.526289e-01
#> gene97  5.549315e-01
#> gene13  5.800327e-01
#> gene15  5.800327e-01
#> gene71  5.800327e-01
#> gene16  5.862913e-01
#> gene28  5.862913e-01
#> gene57  5.862913e-01
#> gene32  6.017309e-01
#> gene96  6.257559e-01
#> gene60  7.016349e-01
#> gene81  7.131380e-01
#> gene99  7.131380e-01
#> gene69  7.378029e-01
#> gene77  7.533213e-01
#> gene43  7.684486e-01
#> gene54  7.684486e-01
#> gene12  7.752865e-01
#> gene27  7.752865e-01
#> gene25  7.939906e-01
#> gene31  7.939906e-01
#> gene62  8.010328e-01
#> gene98  8.010328e-01
#> gene79  8.363295e-01
#> gene8   8.522053e-01
#> gene84  8.828822e-01
#> gene91  8.828822e-01
#> gene74  8.966815e-01
#> gene82  8.966815e-01
#> gene85  9.349298e-01
#> gene18  9.441468e-01
#> gene50  9.441468e-01
#> gene21  9.447563e-01
#> gene61  9.565496e-01
#> gene67  9.565496e-01
#> gene94  9.567623e-01
#> gene3             NA
#> gene6             NA
#> gene7             NA
#> gene17            NA
#> gene19            NA
#> gene22            NA
#> gene26            NA
#> gene35            NA
#> gene38            NA
#> gene42            NA
#> gene44            NA
#> gene45            NA
#> gene47            NA
#> gene53            NA
#> gene55            NA
#> gene59            NA
#> gene70            NA
#> gene73            NA
#> gene83            NA
#> gene86            NA
#> gene87            NA
#> gene90            NA