Read in a GMT file

Returns a list of pathways from a GMT file.

read_gmt(gmtfile)

Arguments

gmtfile
A character value, containing the location of the GMT formatted file. It can also be a file found online

Value

A list of vectors, one for each pathway in the GMT file.

Examples

# this example reads in the freely available pathways from wikipathways
mysigs <- read_gmt(
  "http://data.wikipathways.org/20180910/gmt/wikipathways-20180910-gmt-Homo_sapiens.gmt")
head(mysigs)

#> $`B Cell Receptor Signaling Pathway%WikiPathways_20180910%WP23%Homo sapiens`
#>  [1] "4690"      "5781"      "11184"     "6195"      "4609"      "998"      
#>  [7] "5777"      "10892"     "3635"      "4208"      "2308"      "4209"     
#> [13] "2549"      "4067"      "6885"      "640"       "118788"    "84433"    
#> [19] "5788"      "7462"      "6654"      "3662"      "3265"      "8394"     
#> [25] "8395"      "29760"     "1871"      "930"       "933"       "815"      
#> [31] "3932"      "2969"      "5294"      "5295"      "5296"      "5970"     
#> [37] "3551"      "5170"      "5608"      "23396"     "5601"      "5966"     
#> [43] "5604"      "5605"      "27071"     "5580"      "7006"      "4773"     
#> [49] "1385"      "1380"      "673"       "25780"     "5335"      "5336"     
#> [55] "5579"      "100847070" "8462"      "5594"      "7410"      "5596"     
#> [61] "1147"      "2113"      "1386"      "4775"      "207"       "604"      
#> [67] "3608"      "7409"      "3725"      "4790"      "6452"      "5880"     
#> [73] "6850"      "4792"      "2889"      "695"       "2885"      "1432"     
#> [79] "973"       "1399"      "3059"      "974"       "1398"      "2002"     
#> [85] "975"       "5599"      "4149"      "5879"      "6464"      "5894"     
#> [91] "26469"     "2932"      "2931"      "2534"      "8517"      "867"      
#> [97] "9846"      "8915"     
#> 
#> $`Trans-sulfuration pathway%WikiPathways_20180910%WP2333%Homo sapiens`
#>  [1] "1786"  "2730"  "27430" "875"   "191"   "4357"  "1491"  "2805"  "4548" 
#> [10] "51380"
#> 
#> $`Nanoparticle triggered autophagic cell death%WikiPathways_20180910%WP2509%Homo sapiens`
#>  [1] "9474"   "8678"   "7248"   "285973" "64422"  "55054"  "22863"  "406991"
#>  [9] "9140"   "83734"  "596"    "51100"  "84557"  "115201" "3630"   "55626" 
#> [17] "7249"   "3643"   "81671"  "7405"   "10036"  "8408"   "9706"   "10533" 
#> 
#> $`Benzene metabolism%WikiPathways_20180910%WP3891%Homo sapiens`
#> [1] "2944"  "27294" "4353"  "1571"  "1728"  "2052" 
#> 
#> $`Codeine and Morphine Metabolism%WikiPathways_20180910%WP1604%Homo sapiens`
#>  [1] "7363"  "7364"  "5243"  "1558"  "1576"  "1565"  "54578" "54600" "1244" 
#> [10] "54576" "54575" "8714"  "54659" "54658" "10599"
#> 
#> $`Calcium Regulation in the Cardiac Cell%WikiPathways_20180910%WP536%Homo sapiens`
#>   [1] "102465524" "3760"      "8490"      "2792"      "2791"      "2702"     
#>   [7] "111"       "2701"      "2700"      "2788"      "112"       "2787"     
#>  [13] "113"       "114"       "2786"      "115"       "2785"      "2784"     
#>  [19] "127534"    "2783"      "51764"     "2709"      "2707"      "2706"     
#>  [25] "2705"      "2703"      "6001"      "94235"     "6000"      "8787"     
#>  [31] "406974"    "6004"      "8786"      "480"       "481"       "482"      
#>  [37] "2280"      "483"       "55970"     "486"       "488"       "489"      
#>  [43] "3762"      "801"       "9628"      "8536"      "408"       "805"      
#>  [49] "409"       "808"       "490"       "491"       "492"       "57369"    
#>  [55] "10636"     "773"       "774"       "775"       "776"       "777"      
#>  [61] "811"       "779"       "814"       "815"       "816"       "817"      
#>  [67] "818"       "6263"      "6262"      "5570"      "5331"      "5573"     
#>  [73] "8601"      "5575"      "5576"      "1132"      "1131"      "6261"     
#>  [79] "1129"      "782"       "1128"      "146"       "784"       "3708"     
#>  [85] "10804"     "147"       "148"       "2697"      "3709"      "5566"     
#>  [91] "5567"      "5569"      "309"       "5580"      "5581"      "5582"     
#>  [97] "5583"      "11142"     "10052"     "5587"      "2870"      "153"      
#> [103] "154"       "155"       "1133"      "6546"      "5577"      "5578"     
#> [109] "5579"      "3710"      "6547"      "10971"     "2869"      "2868"     
#> [115] "5350"      "7532"      "7531"      "7534"      "7533"      "64407"    
#> [121] "5590"      "844"       "5588"      "845"       "7529"      "196883"   
#> [127] "2771"      "2770"      "57165"     "54331"     "2767"      "59345"    
#> [133] "26575"     "5996"      "5997"      "5998"      "5999"      "2782"     
#> [139] "2781"      "2778"      "2810"      "2776"      "2775"      "81025"    
#> [145] "2773"      "107"       "108"       "10681"     "109"       "10287"    
#> [151] "3745"     
#> 

# see how the gene identifiers are encoded as ENTREZ id