Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult

A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n = 100, m = 8, betaSD = 1)
dds <- DESeq2::DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- DESeq2::results(dds)
deseqresult2tbl(res)
#>              id     baseMean log2FoldChange     lfcSE         stat       pvalue
#> gene30   gene30  270.8368993    3.727703707 0.3987673  9.348067842 8.926341e-21
#> gene84   gene84  368.6603814    2.730850718 0.4151773  6.577552648 4.782545e-11
#> gene79   gene79  189.7759347    2.178867869 0.3725984  5.847764730 4.982225e-09
#> gene87   gene87  127.5593826    2.585009462 0.4447775  5.811916350 6.176170e-09
#> gene99   gene99   62.6699251   -1.735962993 0.4169444 -4.163535510 3.133573e-05
#> gene31   gene31   39.5372967    1.580768884 0.4459083  3.545053751 3.925330e-04
#> gene36   gene36 1141.2845247    1.369185852 0.3865841  3.541753941 3.974761e-04
#> gene59   gene59   64.9463963   -1.246477285 0.3628867 -3.434894045 5.927856e-04
#> gene55   gene55   20.0460161    2.057324356 0.6152909  3.343661402 8.268058e-04
#> gene61   gene61   59.6541003   -1.422286553 0.4341062 -3.276356413 1.051558e-03
#> gene49   gene49   37.3894312    1.790903876 0.5635788  3.177734757 1.484305e-03
#> gene15   gene15   65.9077303   -1.149321527 0.3667476 -3.133821536 1.725458e-03
#> gene98   gene98   26.6836439    1.772484182 0.5831560  3.039468429 2.369960e-03
#> gene16   gene16  158.1383397   -0.830335635 0.2985051 -2.781646040 5.408399e-03
#> gene23   gene23 1324.8749281   -0.835229621 0.3041520 -2.746092936 6.030968e-03
#> gene100 gene100    3.9143135   -3.935037425 1.4285109 -2.754642823 5.875624e-03
#> gene96   gene96   40.4950807    1.188471292 0.4553039  2.610281558 9.046773e-03
#> gene19   gene19   42.7210421   -1.285418720 0.5364958 -2.395953018 1.657722e-02
#> gene70   gene70   22.0657774   -1.512952572 0.6754394 -2.239953036 2.509397e-02
#> gene83   gene83   29.8464216   -1.212369494 0.5631426 -2.152864101 3.132936e-02
#> gene18   gene18   17.3944162    1.691143769 0.8014944  2.109988248 3.485937e-02
#> gene33   gene33   10.3831582    1.671404067 0.7966178  2.098125371 3.589407e-02
#> gene80   gene80   26.1697348   -1.169565090 0.5693177 -2.054327488 3.994401e-02
#> gene41   gene41  196.6532070    0.755686362 0.3753496  2.013286816 4.408446e-02
#> gene32   gene32  107.3147801    0.806801722 0.4064410  1.985040335 4.713998e-02
#> gene52   gene52   22.0803833   -1.220663599 0.6272336 -1.946106948 5.164189e-02
#> gene76   gene76   27.6556127   -0.938138467 0.4799812 -1.954531616 5.063838e-02
#> gene60   gene60   18.6018243   -1.040441345 0.5770626 -1.802995748 7.138888e-02
#> gene9     gene9   41.7546118   -0.786856699 0.4528497 -1.737566854 8.228716e-02
#> gene85   gene85   10.7724968    1.398888365 0.8219607  1.701891932 8.877563e-02
#> gene88   gene88   13.8546396   -1.168970891 0.6688648 -1.747693908 8.051704e-02
#> gene89   gene89   71.2021337   -0.758769184 0.4402284 -1.723580646 8.478357e-02
#> gene91   gene91   34.3636649   -0.747046095 0.4388087 -1.702441342 8.867267e-02
#> gene94   gene94  350.6455394   -0.708592023 0.4078790 -1.737260531 8.234119e-02
#> gene75   gene75    2.1495441    2.945699384 1.7610934  1.672653725 9.439548e-02
#> gene13   gene13   19.8550785    1.072152099 0.6535410  1.640527714 1.008955e-01
#> gene25   gene25   13.1139906    1.057403778 0.6404299  1.651084462 9.872133e-02
#> gene6     gene6  743.1116666   -0.478248034 0.3147477 -1.519464898 1.286455e-01
#> gene27   gene27   31.6057545    0.829805259 0.5518028  1.503807597 1.326309e-01
#> gene43   gene43   23.4522434   -0.854026984 0.5642870 -1.513462074 1.301623e-01
#> gene35   gene35    9.7519199    1.337172900 0.9185783  1.455698259 1.454761e-01
#> gene22   gene22    6.0609300    1.510336017 1.0776867  1.401461162 1.610762e-01
#> gene26   gene26    9.0471217    1.148002506 0.8285177  1.385610183 1.658660e-01
#> gene47   gene47   19.5391200   -0.842501717 0.6230376 -1.352248577 1.762958e-01
#> gene2     gene2    5.2016038   -1.396027758 1.0758675 -1.297583301 1.944306e-01
#> gene4     gene4   43.9179963   -0.661554967 0.5094771 -1.298497868 1.941163e-01
#> gene34   gene34    5.5937557   -1.226184636 0.9557521 -1.282952609 1.995087e-01
#> gene28   gene28   49.5266094    1.081997678 0.8673407  1.247488595 2.122184e-01
#> gene29   gene29    7.5463437   -1.020036972 0.8081935 -1.262119734 2.069057e-01
#> gene63   gene63   21.5624750   -0.697404912 0.5619981 -1.240938224 2.146286e-01
#> gene82   gene82   23.1305916   -0.794032835 0.6678099 -1.189010329 2.344356e-01
#> gene21   gene21  154.2435300    0.416331737 0.3689617  1.128387417 2.591563e-01
#> gene73   gene73    8.1776913    0.952548458 0.8379636  1.136742084 2.556461e-01
#> gene77   gene77 1024.1377473   -0.322876211 0.2866438 -1.126402157 2.599953e-01
#> gene7     gene7    4.5881439   -1.142745684 1.0883818 -1.049949311 2.937414e-01
#> gene24   gene24   53.1914445    0.530217451 0.5202760  1.019107963 3.081517e-01
#> gene38   gene38   45.0660878   -0.483064713 0.4716677 -1.024163219 3.057582e-01
#> gene53   gene53   12.1564516    0.750761779 0.7441293  1.008913052 3.130163e-01
#> gene92   gene92   14.5534238   -0.686601230 0.6754582 -1.016497052 3.093928e-01
#> gene65   gene65    9.3228199   -0.750413028 0.7809860 -0.960853369 3.366259e-01
#> gene86   gene86   20.6528719   -0.629781174 0.6596505 -0.954719473 3.397196e-01
#> gene69   gene69    0.9334287   -2.496013536 2.6895013 -0.928058141 3.533774e-01
#> gene40   gene40   16.1697595    0.503306292 0.5771964  0.871984485 3.832168e-01
#> gene51   gene51  124.0074864    0.352419576 0.4133578  0.852577452 3.938937e-01
#> gene8     gene8   63.1949465    0.334644856 0.4479650  0.747033421 4.550434e-01
#> gene44   gene44    0.5312864   -2.606002366 3.4517020 -0.754990549 4.502547e-01
#> gene66   gene66    0.4758616    2.431320764 3.3293164  0.730276269 4.652213e-01
#> gene93   gene93  214.0066250   -0.302155497 0.4091919 -0.738420067 4.602592e-01
#> gene95   gene95   58.4521995    0.390854502 0.5241561  0.745683422 4.558587e-01
#> gene3     gene3  107.2479305    0.253302234 0.3954200  0.640590403 5.217888e-01
#> gene1     gene1    5.0398924   -0.530305140 0.9409192 -0.563603301 5.730241e-01
#> gene46   gene46    5.2917846    1.226865233 2.0792478  0.590052428 5.551555e-01
#> gene58   gene58    2.1237693   -0.962478898 1.6997102 -0.566260584 5.712167e-01
#> gene64   gene64    8.6314649    0.494807923 0.8441270  0.586177126 5.577565e-01
#> gene50   gene50    0.3187718   -1.877061565 3.4958541 -0.536939337 5.913095e-01
#> gene97   gene97    7.1687089    0.470278325 0.8809418  0.533835866 5.934551e-01
#> gene68   gene68   10.2303567   -0.332506108 0.7146053 -0.465300366 6.417164e-01
#> gene71   gene71   44.0786866   -0.214900196 0.4945530 -0.434534183 6.639006e-01
#> gene20   gene20    6.8784544    0.400581753 0.9536994  0.420029379 6.744640e-01
#> gene12   gene12   35.8111518    0.200374268 0.4930600  0.406389227 6.844566e-01
#> gene17   gene17    1.4907035    0.743329137 1.9580185  0.379633365 7.042176e-01
#> gene62   gene62   18.2026895   -0.210281154 0.5855503 -0.359117168 7.195074e-01
#> gene11   gene11    9.8290465   -0.227065218 0.7833645 -0.289858969 7.719241e-01
#> gene48   gene48   89.1075961   -0.124350672 0.4599763 -0.270341477 7.868976e-01
#> gene56   gene56    9.0851147   -0.219308138 0.8374575 -0.261873765 7.934188e-01
#> gene57   gene57   30.2203885   -0.134775201 0.5203266 -0.259020371 7.956195e-01
#> gene81   gene81    2.4787762    0.417384267 1.3928846  0.299654602 7.644406e-01
#> gene90   gene90   13.5111797    0.202311929 0.6535020  0.309581171 7.568795e-01
#> gene10   gene10   67.2363070    0.098327426 0.4206453  0.233753778 8.151761e-01
#> gene39   gene39   41.2091267    0.125801706 0.5237344  0.240201329 8.101742e-01
#> gene37   gene37    0.3491292    0.701921072 3.5331283  0.198668434 8.425221e-01
#> gene72   gene72    0.1184641    0.718621307 3.5338115  0.203355866 8.388569e-01
#> gene54   gene54   89.3191312    0.071351604 0.3829045  0.186343092 8.521757e-01
#> gene42   gene42    7.2720475    0.131967440 0.9212457  0.143248911 8.860936e-01
#> gene67   gene67   57.6963520   -0.050981057 0.4356304 -0.117028238 9.068377e-01
#> gene5     gene5  661.2078328   -0.002134634 0.4115102 -0.005187317 9.958611e-01
#> gene45   gene45    0.4746321    0.005943176 3.1575177  0.001882230 9.984982e-01
#> gene78   gene78    9.6382953    0.011913512 1.0244983  0.011628630 9.907219e-01
#> gene14   gene14    0.0000000             NA        NA           NA           NA
#> gene74   gene74    0.0000000             NA        NA           NA           NA
#>                 padj
#> gene30  8.747814e-19
#> gene84  2.343447e-09
#> gene79  1.513162e-07
#> gene87  1.513162e-07
#> gene99  6.141802e-04
#> gene31  5.564665e-03
#> gene36  5.564665e-03
#> gene59  7.261623e-03
#> gene55  9.002997e-03
#> gene61  1.030527e-02
#> gene49  1.322381e-02
#> gene15  1.409124e-02
#> gene98  1.786586e-02
#> gene16  3.693968e-02
#> gene23  3.693968e-02
#> gene100 3.693968e-02
#> gene96  5.215198e-02
#> gene19  9.025373e-02
#> gene70  1.294321e-01
#> gene83  1.535139e-01
#> gene18  1.598918e-01
#> gene33  1.598918e-01
#> gene80  1.701962e-01
#> gene41  1.800116e-01
#> gene32  1.847887e-01
#> gene52  1.874409e-01
#> gene76  1.874409e-01
#> gene60  2.498611e-01
#> gene9   2.558827e-01
#> gene85  2.558827e-01
#> gene88  2.558827e-01
#> gene89  2.558827e-01
#> gene91  2.558827e-01
#> gene94  2.558827e-01
#> gene75  2.643074e-01
#> gene13  2.672367e-01
#> gene25  2.672367e-01
#> gene6   3.249457e-01
#> gene27  3.249457e-01
#> gene43  3.249457e-01
#> gene35  3.477232e-01
#> gene22  3.758445e-01
#> gene26  3.780201e-01
#> gene47  3.926588e-01
#> gene2   4.142216e-01
#> gene4   4.142216e-01
#> gene34  4.159968e-01
#> gene28  4.206720e-01
#> gene29  4.206720e-01
#> gene63  4.206720e-01
#> gene82  4.504841e-01
#> gene21  4.718434e-01
#> gene73  4.718434e-01
#> gene77  4.718434e-01
#> gene7   5.199254e-01
#> gene24  5.199254e-01
#> gene38  5.199254e-01
#> gene53  5.199254e-01
#> gene92  5.199254e-01
#> gene65  5.457790e-01
#> gene86  5.457790e-01
#> gene69  5.585643e-01
#> gene40  5.961151e-01
#> gene51  6.031497e-01
#> gene8   6.607491e-01
#> gene44  6.607491e-01
#> gene66  6.607491e-01
#> gene93  6.607491e-01
#> gene95  6.607491e-01
#> gene3   7.305044e-01
#> gene1   7.588698e-01
#> gene46  7.588698e-01
#> gene58  7.588698e-01
#> gene64  7.588698e-01
#> gene50  7.652447e-01
#> gene97  7.652447e-01
#> gene68  8.167299e-01
#> gene71  8.341315e-01
#> gene20  8.366769e-01
#> gene12  8.384594e-01
#> gene17  8.520163e-01
#> gene62  8.598991e-01
#> gene11  8.860308e-01
#> gene48  8.860308e-01
#> gene56  8.860308e-01
#> gene57  8.860308e-01
#> gene81  8.860308e-01
#> gene90  8.860308e-01
#> gene10  8.876362e-01
#> gene39  8.876362e-01
#> gene37  8.974692e-01
#> gene72  8.974692e-01
#> gene54  8.979916e-01
#> gene42  9.237997e-01
#> gene67  9.354747e-01
#> gene5   9.984982e-01
#> gene45  9.984982e-01
#> gene78  9.984982e-01
#> gene14            NA
#> gene74            NA