Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult

A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n = 100, m = 8, betaSD = 1)
dds <- DESeq2::DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- DESeq2::results(dds)
deseqresult2tbl(res)
#>              id    baseMean log2FoldChange     lfcSE        stat       pvalue
#> gene64   gene64  75.5746595     3.11624143 0.5148624  6.05257134 1.425518e-09
#> gene20   gene20  74.3796071    -1.64664707 0.3315658 -4.96627477 6.825123e-07
#> gene38   gene38 112.7688963     1.51182829 0.3227003  4.68492989 2.800558e-06
#> gene70   gene70  90.0091964     1.88694123 0.4067402  4.63918025 3.497939e-06
#> gene3     gene3 175.6252012     1.57414874 0.3629079  4.33759798 1.440483e-05
#> gene50   gene50  32.9661356     2.24519422 0.5158356  4.35253806 1.345705e-05
#> gene80   gene80  75.6116105    -1.43374098 0.3662699 -3.91443803 9.061501e-05
#> gene46   gene46  42.5471618     1.94044408 0.5057823  3.83652029 1.247899e-04
#> gene99   gene99 246.5191186     1.68843866 0.4387844  3.84799185 1.190900e-04
#> gene52   gene52  50.7012636    -1.63159432 0.4294624 -3.79915512 1.451902e-04
#> gene90   gene90  54.7923567     2.04184568 0.5639080  3.62088439 2.935977e-04
#> gene13   gene13  80.5098262     1.67302653 0.4794070  3.48978308 4.834128e-04
#> gene10   gene10  27.7391581     2.34707927 0.7076890  3.31654066 9.113932e-04
#> gene85   gene85  40.0666063    -1.31128031 0.4353136 -3.01226593 2.593053e-03
#> gene72   gene72  19.2144934    -2.32284662 0.7863406 -2.95399565 3.136884e-03
#> gene42   gene42  48.5928491     1.46470680 0.5061497  2.89382138 3.805846e-03
#> gene15   gene15  79.0299506    -1.27775911 0.4903870 -2.60561355 9.170987e-03
#> gene23   gene23 185.3757172     0.79044163 0.3032561  2.60651540 9.146871e-03
#> gene75   gene75  20.8288187     1.16397547 0.4954114  2.34951306 1.879798e-02
#> gene78   gene78   9.0475029    -1.54556876 0.6686461 -2.31148994 2.080581e-02
#> gene93   gene93   4.4812281    -3.18468007 1.3912743 -2.28903820 2.207713e-02
#> gene87   gene87  33.9215343     1.01909204 0.4828245  2.11068829 3.479911e-02
#> gene96   gene96  39.9522206    -0.99741352 0.4720104 -2.11311750 3.459071e-02
#> gene1     gene1  57.4740354    -0.65567345 0.3287940 -1.99417717 4.613270e-02
#> gene32   gene32  11.8724190    -1.54916739 0.7837025 -1.97672888 4.807228e-02
#> gene33   gene33  10.7536159    -1.23825889 0.6588827 -1.87933130 6.019927e-02
#> gene35   gene35  10.2151621    -1.26930416 0.6697640 -1.89515127 5.807236e-02
#> gene40   gene40  15.3611159    -1.40774270 0.7451106 -1.88930708 5.885069e-02
#> gene49   gene49   6.4105708     2.03359519 1.0638704  1.91150647 5.593953e-02
#> gene24   gene24 149.3180103     0.55551278 0.2990077  1.85785423 6.318972e-02
#> gene61   gene61  98.4660663    -0.72246362 0.4051320 -1.78327965 7.454079e-02
#> gene73   gene73   1.0030855    -3.62084283 2.0358805 -1.77851445 7.531940e-02
#> gene77   gene77  23.2470489     1.41802401 0.8342742  1.69970970 8.918554e-02
#> gene45   gene45  31.2806316     0.93181575 0.5624991  1.65656385 9.760769e-02
#> gene37   gene37  23.2591011    -0.82165576 0.5223346 -1.57304478 1.157085e-01
#> gene41   gene41   8.9947156     1.15228437 0.7331354  1.57172114 1.160152e-01
#> gene56   gene56   7.7195963    -1.55925355 0.9871243 -1.57959192 1.142003e-01
#> gene68   gene68  10.1988942     1.30574391 0.8039967  1.62406632 1.043617e-01
#> gene97   gene97  16.7417612    -1.09495213 0.6918457 -1.58265377 1.135004e-01
#> gene31   gene31   2.3135849     2.25708799 1.4733998  1.53189108 1.255493e-01
#> gene5     gene5  10.4738553    -0.95314499 0.6625013 -1.43870661 1.502337e-01
#> gene22   gene22   4.6225639    -1.90972751 1.3986864 -1.36537222 1.721361e-01
#> gene36   gene36   9.6244973    -1.07464257 0.7944036 -1.35276649 1.761302e-01
#> gene74   gene74  14.9077861    -0.79650669 0.5851835 -1.36112299 1.734748e-01
#> gene14   gene14  20.7169880    -0.73383971 0.5607552 -1.30866325 1.906484e-01
#> gene63   gene63  23.1247192    -0.73103185 0.5630684 -1.29830032 1.941842e-01
#> gene11   gene11   4.6802220    -1.31075755 1.0609144 -1.23549794 2.166452e-01
#> gene92   gene92   5.3743518     1.10028762 0.9308287  1.18205165 2.371852e-01
#> gene44   gene44  24.2803612    -0.51420089 0.4602246 -1.11728241 2.638736e-01
#> gene62   gene62   3.7243062     1.13140989 1.0047168  1.12609830 2.601239e-01
#> gene28   gene28 100.6005904    -0.53898688 0.5053442 -1.06657373 2.861644e-01
#> gene51   gene51   2.6520841    -1.27423737 1.2319055 -1.03436291 3.009665e-01
#> gene60   gene60  21.2771943    -0.60769064 0.5914483 -1.02746190 3.042030e-01
#> gene16   gene16 355.8655894     0.35297224 0.3493746  1.01029732 3.123529e-01
#> gene81   gene81   2.3459190    -1.74673860 1.7671242 -0.98846397 3.229255e-01
#> gene25   gene25   2.9915813    -1.27180805 1.3108211 -0.97023771 3.319280e-01
#> gene4     gene4   3.2380639     1.17962000 1.2781830  0.92288818 3.560655e-01
#> gene26   gene26   5.2400837     1.13728013 1.1959273  0.95096095 3.416242e-01
#> gene66   gene66  46.9185327     0.46395514 0.4959889  0.93541429 3.495748e-01
#> gene82   gene82  17.6613344     0.52424675 0.5664901  0.92542968 3.547424e-01
#> gene21   gene21   6.5223592     0.87211601 1.0033384  0.86921422 3.847300e-01
#> gene43   gene43   8.9357015    -0.72201723 0.8382575 -0.86133112 3.890557e-01
#> gene12   gene12  33.3786155    -0.39009997 0.4612069 -0.84582421 3.976508e-01
#> gene98   gene98  20.0409812    -0.51110086 0.6153619 -0.83056949 4.062169e-01
#> gene55   gene55  49.8158136     0.49842035 0.6324275  0.78810672 4.306343e-01
#> gene67   gene67  13.9490197    -0.65595682 0.8929204 -0.73461959 4.625712e-01
#> gene100 gene100  32.1853180     0.31616342 0.4379426  0.72192886 4.703382e-01
#> gene65   gene65  30.4798646    -0.44281180 0.6272256 -0.70598493 4.801975e-01
#> gene7     gene7   6.0180962     0.61882223 0.9110971  0.67920557 4.970076e-01
#> gene18   gene18  24.0381383     0.35621591 0.5211342  0.68353971 4.942659e-01
#> gene88   gene88  14.6216993     0.50067237 0.9212818  0.54345193 5.868187e-01
#> gene53   gene53  44.4569521    -0.28356863 0.5696714 -0.49777579 6.186421e-01
#> gene17   gene17   2.8018099    -0.56511132 1.2930391 -0.43704117 6.620815e-01
#> gene59   gene59 127.3070324     0.17585661 0.3689140  0.47668723 6.335849e-01
#> gene79   gene79   6.3470895    -0.41082117 0.9239484 -0.44463648 6.565825e-01
#> gene89   gene89   5.8937279     0.38264121 0.8230585  0.46490158 6.420019e-01
#> gene91   gene91   5.1895988    -0.44756631 0.9713779 -0.46075406 6.449751e-01
#> gene19   gene19  38.7376812    -0.23714119 0.5948304 -0.39867028 6.901362e-01
#> gene57   gene57  11.2511327     0.29579320 0.7829688  0.37778413 7.055910e-01
#> gene58   gene58   8.7507610    -0.37442633 0.9809936 -0.38168071 7.026982e-01
#> gene9     gene9   9.3790896    -0.22963971 0.7571115 -0.30331031 7.616534e-01
#> gene47   gene47  10.0344774    -0.21796315 0.7664335 -0.28438627 7.761144e-01
#> gene84   gene84  70.5812774     0.15389770 0.5770131  0.26671438 7.896891e-01
#> gene86   gene86  22.2012494     0.23800490 0.8757891  0.27176052 7.858062e-01
#> gene54   gene54   1.2569180    -0.39840385 1.8656533 -0.21354656 8.309007e-01
#> gene69   gene69  20.8215750    -0.11415514 0.5050994 -0.22600528 8.211973e-01
#> gene83   gene83  18.6201576    -0.15288794 0.6899695 -0.22158650 8.246358e-01
#> gene94   gene94   9.2251433     0.13625708 0.6743030  0.20207101 8.398612e-01
#> gene2     gene2   2.4541540     0.30890831 1.7930410  0.17228179 8.632160e-01
#> gene29   gene29  14.9136708    -0.09748638 0.5857665 -0.16642531 8.678223e-01
#> gene6     gene6   3.6300282    -0.16148446 1.0826255 -0.14916004 8.814274e-01
#> gene27   gene27  21.2764320     0.06608363 0.5981103  0.11048736 9.120229e-01
#> gene48   gene48  17.2098923     0.05883816 0.5992042  0.09819384 9.217784e-01
#> gene39   gene39   0.9882897    -0.14373931 2.1880177 -0.06569385 9.476216e-01
#> gene8     gene8   0.7347302    -0.08755454 2.9924289 -0.02925869 9.766583e-01
#> gene30   gene30   0.2372382    -0.19745452 3.5338115 -0.05587579 9.554408e-01
#> gene34   gene34   0.4524821     2.26937142 3.4582376  0.65622196 5.116813e-01
#> gene71   gene71   0.4205203     2.17343310 3.4633249  0.62755680 5.302943e-01
#> gene76   gene76   0.8911131     0.11505771 2.5133787  0.04577810 9.634871e-01
#> gene95   gene95   0.8656068     1.35810174 2.8729959  0.47271273 6.364181e-01
#>                 padj
#> gene64  1.339987e-07
#> gene20  3.207808e-05
#> gene38  8.220156e-05
#> gene70  8.220156e-05
#> gene3   2.256757e-04
#> gene50  2.256757e-04
#> gene80  1.216830e-03
#> gene46  1.303361e-03
#> gene99  1.303361e-03
#> gene52  1.364788e-03
#> gene90  2.508926e-03
#> gene13  3.786733e-03
#> gene10  6.590074e-03
#> gene85  1.741050e-02
#> gene72  1.965781e-02
#> gene42  2.235934e-02
#> gene15  4.789293e-02
#> gene23  4.789293e-02
#> gene75  9.300056e-02
#> gene78  9.778730e-02
#> gene93  9.882145e-02
#> gene87  1.422225e-01
#> gene96  1.422225e-01
#> gene1   1.806864e-01
#> gene32  1.807518e-01
#> gene33  1.951287e-01
#> gene35  1.951287e-01
#> gene40  1.951287e-01
#> gene49  1.951287e-01
#> gene24  1.979944e-01
#> gene61  2.212507e-01
#> gene73  2.212507e-01
#> gene77  2.540437e-01
#> gene45  2.698565e-01
#> gene37  2.796265e-01
#> gene41  2.796265e-01
#> gene56  2.796265e-01
#> gene68  2.796265e-01
#> gene97  2.796265e-01
#> gene31  2.950409e-01
#> gene5   3.444382e-01
#> gene22  3.762782e-01
#> gene36  3.762782e-01
#> gene74  3.762782e-01
#> gene14  3.968111e-01
#> gene63  3.968111e-01
#> gene11  4.332905e-01
#> gene92  4.644877e-01
#> gene44  4.960824e-01
#> gene62  4.960824e-01
#> gene28  5.274402e-01
#> gene51  5.395299e-01
#> gene60  5.395299e-01
#> gene16  5.437254e-01
#> gene81  5.519090e-01
#> gene25  5.571649e-01
#> gene4   5.578359e-01
#> gene26  5.578359e-01
#> gene66  5.578359e-01
#> gene82  5.578359e-01
#> gene21  5.898586e-01
#> gene43  5.898586e-01
#> gene12  5.933203e-01
#> gene98  5.966310e-01
#> gene55  6.227634e-01
#> gene67  6.588135e-01
#> gene100 6.598775e-01
#> gene65  6.638024e-01
#> gene7   6.674102e-01
#> gene18  6.674102e-01
#> gene88  7.769149e-01
#> gene53  8.076716e-01
#> gene17  8.082553e-01
#> gene59  8.082553e-01
#> gene79  8.082553e-01
#> gene89  8.082553e-01
#> gene91  8.082553e-01
#> gene19  8.290694e-01
#> gene57  8.290694e-01
#> gene58  8.290694e-01
#> gene9   8.836997e-01
#> gene47  8.836997e-01
#> gene84  8.836997e-01
#> gene86  8.836997e-01
#> gene54  8.971245e-01
#> gene69  8.971245e-01
#> gene83  8.971245e-01
#> gene94  8.971245e-01
#> gene2   9.063921e-01
#> gene29  9.063921e-01
#> gene6   9.104854e-01
#> gene27  9.316900e-01
#> gene48  9.316900e-01
#> gene39  9.476216e-01
#> gene8             NA
#> gene30            NA
#> gene34            NA
#> gene71            NA
#> gene76            NA
#> gene95            NA