Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult
A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n=100, m=8, betaSD = 1)
dds <- DESeq2::DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- DESeq2::results(dds)
deseqresult2tbl(res)

#>          id     baseMean log2FoldChange     lfcSE        stat       pvalue
#> 1    gene36  201.0507359    -1.91637009 0.3580369 -5.35243727 8.677739e-08
#> 2    gene59   38.2390390    -1.97490148 0.4625443 -4.26964850 1.957813e-05
#> 3    gene17  130.9482530     1.81280585 0.4503252  4.02554792 5.684280e-05
#> 4    gene45   13.6408634    -3.51065594 0.9231292 -3.80299503 1.429572e-04
#> 5    gene35   96.6934860     1.59678774 0.4284141  3.72720640 1.936139e-04
#> 6    gene22   11.9918964     2.45245455 0.6991079  3.50797734 4.515275e-04
#> 7    gene30   20.0360952     2.68369641 0.7632626  3.51608521 4.379604e-04
#> 8    gene47   30.9990213     2.30200572 0.6639188  3.46730003 5.257147e-04
#> 9    gene48  141.3903991    -1.53763754 0.4616571 -3.33069193 8.663042e-04
#> 10   gene14    9.5379376     3.04444275 0.9796747  3.10760580 1.886095e-03
#> 11   gene41   51.4209185     1.41152531 0.4523024  3.12075574 1.803876e-03
#> 12   gene52   62.9607498     1.57170244 0.4977480  3.15762677 1.590591e-03
#> 13   gene99   14.0555290     2.56292407 0.8963561  2.85926995 4.246173e-03
#> 14   gene91  189.7377862     1.17646894 0.4251865  2.76694771 5.658383e-03
#> 15   gene16 1455.3471654     0.75659132 0.2866799  2.63915043 8.311409e-03
#> 16   gene88   48.2869500    -1.18028609 0.4462157 -2.64510182 8.166634e-03
#> 17   gene25   21.7707593    -1.89941728 0.7610589 -2.49575580 1.256891e-02
#> 18   gene95   26.3907332     1.47787506 0.6020864  2.45458957 1.410455e-02
#> 19   gene42   13.4907984    -1.58021762 0.6639147 -2.38015148 1.730552e-02
#> 20   gene58    9.7481458     2.17142034 0.9270738  2.34223032 1.916888e-02
#> 21   gene37    2.4402953    -4.80011346 2.2062455 -2.17569329 2.957820e-02
#> 22   gene10    8.8103726     1.96158892 0.9313189  2.10624844 3.518278e-02
#> 23   gene28   66.0559002    -1.19971225 0.5653907 -2.12191709 3.384470e-02
#> 24    gene7   17.4066851    -1.52573823 0.7318174 -2.08486205 3.708182e-02
#> 25   gene74  119.2444931     1.07894698 0.5215751  2.06863226 3.858061e-02
#> 26   gene13    5.1436895    -1.96841735 1.0215061 -1.92697557 5.398267e-02
#> 27   gene29   46.6583228     1.12207467 0.5823068  1.92694763 5.398616e-02
#> 28   gene34   30.2100259    -0.94097498 0.4939686 -1.90492889 5.678932e-02
#> 29   gene53    4.9470616    -2.13712589 1.1240157 -1.90133103 5.725867e-02
#> 30   gene73    3.2819728    -3.06330500 1.5765642 -1.94302584 5.201304e-02
#> 31   gene78   46.2194460    -0.96429124 0.4897345 -1.96900834 4.895213e-02
#> 32   gene12    5.6974565    -1.77937666 0.9481905 -1.87660256 6.057259e-02
#> 33   gene39    9.4442326     1.71638605 0.9792519  1.75275241 7.964452e-02
#> 34   gene66   32.1660652    -0.85181425 0.4881441 -1.74500589 8.098384e-02
#> 35    gene1    3.5198054    -2.18593849 1.2991981 -1.68252904 9.246630e-02
#> 36   gene70   12.8406837     1.30540972 0.7740237  1.68652412 9.169489e-02
#> 37   gene15   32.2207917    -0.74577264 0.4643537 -1.60604440 1.082642e-01
#> 38    gene9    1.8781306     2.71495310 1.7228546  1.57584577 1.150614e-01
#> 39   gene80   18.1284900    -0.92702859 0.5999858 -1.54508414 1.223259e-01
#> 40   gene94   22.2126979     0.79565441 0.5244279  1.51718564 1.292198e-01
#> 41   gene84   29.2885375    -0.72526491 0.4856281 -1.49345743 1.353175e-01
#> 42   gene19  162.3336274    -0.68614755 0.4642421 -1.47799498 1.394091e-01
#> 43   gene46   17.4703747     1.25219531 0.8837150  1.41696742 1.564925e-01
#> 44   gene72    7.2713694    -1.54119851 1.0944277 -1.40822329 1.590650e-01
#> 45   gene76    4.3020594     1.56066809 1.1181699  1.39573426 1.627945e-01
#> 46   gene18    1.0501704     3.52552666 2.6950938  1.30812762 1.908300e-01
#> 47   gene27    2.1543767    -2.34369633 1.8335938 -1.27819824 2.011795e-01
#> 48   gene57   59.8031312     0.63824022 0.4988948  1.27930811 2.007886e-01
#> 49   gene40    7.8313661    -1.02664479 0.8111100 -1.26572820 2.056104e-01
#> 50   gene90    2.6606524    -1.84950909 1.5770729 -1.17274802 2.408969e-01
#> 51   gene50   55.6350041     0.51476483 0.4642515  1.10880601 2.675139e-01
#> 52   gene56    8.5158935    -0.94347984 0.9232376 -1.02192528 3.068163e-01
#> 53   gene11   91.8175812    -0.37959439 0.4109880 -0.92361432 3.556872e-01
#> 54   gene31   35.3561300     0.46948309 0.5049796  0.92970702 3.525228e-01
#> 55   gene55   97.0705016     0.36218039 0.3831210  0.94534214 3.444842e-01
#> 56   gene64   30.8028830    -0.57012010 0.5904463 -0.96557488 3.342570e-01
#> 57   gene85  240.7001455     0.27672828 0.2911117  0.95059132 3.418119e-01
#> 58   gene65   12.5397454    -0.58198500 0.7207506 -0.80747068 4.193954e-01
#> 59   gene69    7.0641825     0.72006840 0.8967623  0.80296465 4.219952e-01
#> 60   gene82    5.9859845     0.83753196 1.0445717  0.80179463 4.226718e-01
#> 61   gene75   64.6684386    -0.36227857 0.4836927 -0.74898498 4.538663e-01
#> 62   gene96    1.9806257    -1.28040122 1.7402661 -0.73575023 4.618827e-01
#> 63   gene60   19.8481388    -0.40568845 0.5754758 -0.70496182 4.808340e-01
#> 64   gene67    0.4671662    -2.42603193 3.4637965 -0.70039677 4.836796e-01
#> 65   gene20   44.7678101    -0.28044131 0.4542554 -0.61736488 5.369941e-01
#> 66   gene21   31.1775378    -0.30996917 0.4908745 -0.63146314 5.277377e-01
#> 67   gene33    1.7946255    -1.08942598 1.7762899 -0.61331541 5.396678e-01
#> 68   gene62   10.7274065     0.49968441 0.7980207  0.62615467 5.312135e-01
#> 69   gene97    3.4850980     0.93604314 1.4666021  0.63823933 5.233179e-01
#> 70   gene63    2.4620945    -1.01646827 1.6974065 -0.59883607 5.492822e-01
#> 71   gene93   99.3633152    -0.25236673 0.4393463 -0.57441408 5.656876e-01
#> 72    gene4    5.1842760    -0.56554163 1.0290080 -0.54959889 5.825945e-01
#> 73   gene61    2.7150798    -0.88474603 1.6416019 -0.53895284 5.899194e-01
#> 74   gene23    2.3763171    -0.74804964 1.4755063 -0.50697828 6.121701e-01
#> 75   gene98  273.1498742    -0.22550561 0.4497763 -0.50137287 6.161087e-01
#> 76   gene32   18.4339374    -0.26321104 0.5804050 -0.45349544 6.501920e-01
#> 77   gene68   64.6544827     0.23299189 0.5010202  0.46503487 6.419065e-01
#> 78   gene92   19.8304352     0.27598127 0.6290440  0.43873127 6.608563e-01
#> 79   gene54   13.6528498    -0.25666386 0.6660615 -0.38534562 6.999814e-01
#> 80   gene87   43.2866420     0.24092718 0.6171667  0.39037619 6.962584e-01
#> 81   gene83    0.4620323    -1.29503658 3.4897891 -0.37109308 7.105682e-01
#> 82   gene24   32.3292643    -0.15041433 0.4831214 -0.31133856 7.555433e-01
#> 83   gene81    2.3480230     0.44001260 1.4064901  0.31284443 7.543989e-01
#> 84   gene86    5.0535372     0.33680422 1.0728979  0.31392012 7.535817e-01
#> 85   gene26    3.3210260     0.38519669 1.3351659  0.28850099 7.729633e-01
#> 86    gene2    5.3290538    -0.24847217 1.0486652 -0.23694136 8.127023e-01
#> 87   gene43   64.8478925     0.11172858 0.4643181  0.24062936 8.098424e-01
#> 88   gene79   27.2118880    -0.15325431 0.6315026 -0.24268199 8.082518e-01
#> 89   gene89   24.5090830     0.14503396 0.6847378  0.21180948 8.322557e-01
#> 90   gene44   11.0239965     0.15480107 0.7991384  0.19370995 8.464030e-01
#> 91    gene6   30.2312947    -0.07705647 0.5440808 -0.14162688 8.873747e-01
#> 92   gene51   12.1649692     0.09723730 0.8034686  0.12102191 9.036737e-01
#> 93   gene71   34.6364121     0.07978820 0.6756372  0.11809325 9.059938e-01
#> 94   gene77   11.0947160     0.12968294 0.9772684  0.13269941 8.944311e-01
#> 95  gene100   13.5135729    -0.08408674 0.7321575 -0.11484789 9.085657e-01
#> 96   gene38    9.1674931     0.08039182 0.8468053  0.09493542 9.243661e-01
#> 97    gene8    6.8504984     0.05115846 1.0497426  0.04873429 9.611310e-01
#> 98    gene3   10.5909754     2.75057084 1.2321189  2.23239079           NA
#> 99    gene5    0.0000000             NA        NA          NA           NA
#> 100  gene49    0.0000000             NA        NA          NA           NA
#>             padj
#> 1   8.417407e-06
#> 2   9.495391e-04
#> 3   1.837917e-03
#> 4   3.466713e-03
#> 5   3.756110e-03
#> 6   6.256881e-03
#> 7   6.256881e-03
#> 8   6.374291e-03
#> 9   9.336834e-03
#> 10  1.524593e-02
#> 11  1.524593e-02
#> 12  1.524593e-02
#> 13  3.168298e-02
#> 14  3.920451e-02
#> 15  5.038792e-02
#> 16  5.038792e-02
#> 17  7.171672e-02
#> 18  7.600786e-02
#> 19  8.834925e-02
#> 20  9.296907e-02
#> 21  1.366231e-01
#> 22  1.483795e-01
#> 23  1.483795e-01
#> 24  1.496928e-01
#> 25  1.496928e-01
#> 26  1.791642e-01
#> 27  1.791642e-01
#> 28  1.791642e-01
#> 29  1.791642e-01
#> 30  1.791642e-01
#> 31  1.791642e-01
#> 32  1.836107e-01
#> 33  2.310421e-01
#> 34  2.310421e-01
#> 35  2.491453e-01
#> 36  2.491453e-01
#> 37  2.838276e-01
#> 38  2.937092e-01
#> 39  3.042465e-01
#> 40  3.133581e-01
#> 41  3.201414e-01
#> 42  3.219687e-01
#> 43  3.506659e-01
#> 44  3.506659e-01
#> 45  3.509127e-01
#> 46  4.024024e-01
#> 47  4.065503e-01
#> 48  4.065503e-01
#> 49  4.070247e-01
#> 50  4.673399e-01
#> 51  5.088009e-01
#> 52  5.723303e-01
#> 53  6.052922e-01
#> 54  6.052922e-01
#> 55  6.052922e-01
#> 56  6.052922e-01
#> 57  6.052922e-01
#> 58  6.833194e-01
#> 59  6.833194e-01
#> 60  6.833194e-01
#> 61  7.217218e-01
#> 62  7.226230e-01
#> 63  7.330768e-01
#> 64  7.330768e-01
#> 65  7.586634e-01
#> 66  7.586634e-01
#> 67  7.586634e-01
#> 68  7.586634e-01
#> 69  7.586634e-01
#> 70  7.611482e-01
#> 71  7.728408e-01
#> 72  7.838655e-01
#> 73  7.838655e-01
#> 74  7.968340e-01
#> 75  7.968340e-01
#> 76  8.190731e-01
#> 77  8.190731e-01
#> 78  8.218341e-01
#> 79  8.487274e-01
#> 80  8.487274e-01
#> 81  8.509274e-01
#> 82  8.724726e-01
#> 83  8.724726e-01
#> 84  8.724726e-01
#> 85  8.820875e-01
#> 86  8.958196e-01
#> 87  8.958196e-01
#> 88  8.958196e-01
#> 89  9.070652e-01
#> 90  9.122343e-01
#> 91  9.276934e-01
#> 92  9.276934e-01
#> 93  9.276934e-01
#> 94  9.276934e-01
#> 95  9.276934e-01
#> 96  9.339949e-01
#> 97  9.611310e-01
#> 98            NA
#> 99            NA
#> 100           NA