Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult

A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n = 100, m = 8, betaSD = 1)
dds <- DESeq2::DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- DESeq2::results(dds)
deseqresult2tbl(res)
#>              id     baseMean log2FoldChange     lfcSE        stat       pvalue
#> gene12   gene12 2.990341e+02    2.552872642 0.2773333  9.20506930 3.414508e-20
#> gene87   gene87 2.251237e+02    1.392108759 0.2402709  5.79391266 6.876523e-09
#> gene89   gene89 2.252422e+01   -3.713515565 0.6509105 -5.70510944 1.162684e-08
#> gene83   gene83 7.239124e+01    2.105524086 0.3820961  5.51045621 3.579048e-08
#> gene77   gene77 1.439141e+01   -3.823054855 0.8083543 -4.72942957 2.251515e-06
#> gene82   gene82 1.038480e+02    1.652812402 0.3591683  4.60177704 4.189015e-06
#> gene39   gene39 1.152641e+02    1.268216333 0.3073001  4.12696336 3.675849e-05
#> gene30   gene30 1.133782e+01   -3.111958765 0.8548795 -3.64023107 2.723935e-04
#> gene70   gene70 1.257490e+02    1.225840171 0.3366409  3.64138779 2.711723e-04
#> gene29   gene29 8.111272e+01   -1.187139867 0.3432359 -3.45867030 5.428490e-04
#> gene97   gene97 4.877887e+01   -1.442404428 0.4353967 -3.31285095 9.235018e-04
#> gene88   gene88 4.635017e+01    1.477604174 0.4513894  3.27345808 1.062401e-03
#> gene2     gene2 6.182382e+01    1.418359730 0.4396847  3.22585668 1.255962e-03
#> gene79   gene79 6.948570e+00    3.187376050 1.1576421  2.75333463 5.899156e-03
#> gene61   gene61 2.666504e+01   -1.215955346 0.4517566 -2.69161634 7.110669e-03
#> gene100 gene100 2.617993e+03    0.809625512 0.3096109  2.61497747 8.923334e-03
#> gene45   gene45 4.434260e+01    1.207836019 0.4753549  2.54091413 1.105631e-02
#> gene58   gene58 3.862156e+01    1.066809477 0.4250741  2.50970254 1.208329e-02
#> gene75   gene75 6.371499e+02    0.677531485 0.2695646  2.51342878 1.195639e-02
#> gene81   gene81 2.136416e+01    1.229099067 0.5088511  2.41543932 1.571624e-02
#> gene27   gene27 3.506066e+02    0.573786507 0.2422469  2.36860204 1.785545e-02
#> gene40   gene40 1.002426e+01   -1.811325004 0.7819308 -2.31647732 2.053222e-02
#> gene15   gene15 1.788968e+01   -1.343248868 0.6071699 -2.21231145 2.694515e-02
#> gene32   gene32 1.749860e+02   -0.671342933 0.3038136 -2.20971983 2.712461e-02
#> gene90   gene90 1.596303e+01   -1.347218158 0.6051839 -2.22613013 2.600547e-02
#> gene8     gene8 2.496898e+02   -0.717992716 0.3396330 -2.11402504 3.451313e-02
#> gene42   gene42 4.054879e+01   -0.923649893 0.4394949 -2.10161697 3.558684e-02
#> gene74   gene74 1.547969e+01   -1.337304149 0.6497539 -2.05817022 3.957380e-02
#> gene69   gene69 5.145021e+01   -0.724846012 0.3792008 -1.91150974 5.593911e-02
#> gene28   gene28 1.340528e+02   -0.656458630 0.4024459 -1.63117236 1.028540e-01
#> gene55   gene55 4.577072e+01   -0.647413710 0.3949677 -1.63915594 1.011808e-01
#> gene96   gene96 5.196910e+01    0.669256730 0.4084953  1.63834623 1.013495e-01
#> gene63   gene63 1.293047e+02   -0.552889525 0.3428070 -1.61283039 1.067814e-01
#> gene26   gene26 2.003322e+01    0.948876794 0.5952926  1.59396693 1.109434e-01
#> gene66   gene66 3.287794e+01    0.716543079 0.4656261  1.53888098 1.238334e-01
#> gene93   gene93 6.172044e+00    1.477905493 0.9630005  1.53468811 1.248605e-01
#> gene94   gene94 6.853392e+00    1.822369430 1.1823415  1.54132242 1.232383e-01
#> gene5     gene5 1.020381e+01   -1.059260056 0.7102153 -1.49146339 1.358399e-01
#> gene22   gene22 6.946865e+01   -0.538546312 0.3613569 -1.49034456 1.361337e-01
#> gene9     gene9 1.421470e+02   -0.418995682 0.2930459 -1.42979526 1.527758e-01
#> gene35   gene35 5.537081e+01    0.617327318 0.4414445  1.39842557 1.619853e-01
#> gene36   gene36 1.075614e+02   -0.434836673 0.3106335 -1.39983827 1.615618e-01
#> gene44   gene44 9.160687e+00   -1.020674416 0.7323030 -1.39378708 1.633819e-01
#> gene18   gene18 2.817111e+01   -0.675103468 0.4931510 -1.36895900 1.710121e-01
#> gene17   gene17 2.102075e+01   -0.838729275 0.6325965 -1.32585197 1.848887e-01
#> gene68   gene68 2.620716e+01    0.740525060 0.5646077  1.31157445 1.896638e-01
#> gene13   gene13 3.319036e+01    0.677991429 0.5356316  1.26577939 2.055921e-01
#> gene31   gene31 1.569926e+01   -0.737503689 0.6141715 -1.20081075 2.298246e-01
#> gene95   gene95 6.091133e+00   -1.000174105 0.8513656 -1.17478802 2.400795e-01
#> gene50   gene50 1.256673e+01    0.695600079 0.6387103  1.08906983 2.761231e-01
#> gene47   gene47 2.877806e+01   -0.456446716 0.4389180 -1.03993609 2.983696e-01
#> gene56   gene56 2.419105e+02   -0.292050341 0.2913904 -1.00226490 3.162157e-01
#> gene62   gene62 1.263126e+01   -0.669154253 0.6754227 -0.99071923 3.218227e-01
#> gene10   gene10 1.199995e+01    0.802401703 0.8528383  0.94086027 3.467765e-01
#> gene85   gene85 9.741330e+00   -0.719688402 0.8841424 -0.81399605 4.156472e-01
#> gene99   gene99 7.137574e+00   -0.655599287 0.7939379 -0.82575641 4.089423e-01
#> gene80   gene80 2.634893e+01    0.387490445 0.4909984  0.78918875 4.300017e-01
#> gene64   gene64 1.085429e+02    0.228710681 0.3104765  0.73664398 4.613389e-01
#> gene16   gene16 4.633968e+01    0.306868864 0.4421923  0.69397152 4.877001e-01
#> gene21   gene21 7.028516e+00    0.568548858 0.8265352  0.68787012 4.915346e-01
#> gene51   gene51 8.532073e+00    0.679809138 0.9921769  0.68516926 4.932371e-01
#> gene86   gene86 1.460899e+01   -0.437877100 0.6083218 -0.71981167 4.716410e-01
#> gene43   gene43 1.238294e+01    0.420228457 0.6688406  0.62829384 5.298115e-01
#> gene3     gene3 3.007530e+01    0.337185509 0.5542846  0.60832558 5.429716e-01
#> gene34   gene34 8.620444e+00   -0.412666888 0.7792788 -0.52954977 5.964241e-01
#> gene67   gene67 1.886639e+01   -0.262750944 0.4969558 -0.52872096 5.969990e-01
#> gene73   gene73 5.935772e+00   -0.590407218 1.1714580 -0.50399349 6.142659e-01
#> gene7     gene7 9.087263e+00   -0.353968664 0.8316575 -0.42561832 6.703860e-01
#> gene23   gene23 1.134894e+01   -0.290479603 0.7128291 -0.40750243 6.836390e-01
#> gene78   gene78 2.113269e+01   -0.222627829 0.5740328 -0.38783120 6.981410e-01
#> gene49   gene49 2.312509e+01   -0.174687177 0.4912950 -0.35556472 7.221666e-01
#> gene52   gene52 2.886530e+01   -0.160124589 0.4478098 -0.35757277 7.206631e-01
#> gene37   gene37 7.675249e+01    0.076756130 0.3892151  0.19720748 8.436652e-01
#> gene25   gene25 1.190252e+01    0.103487647 0.6629550  0.15610056 8.759538e-01
#> gene76   gene76 4.536891e+01    0.004561141 0.4437708  0.01027815 9.917994e-01
#> gene1     gene1 4.106427e+00    0.019710570 1.1035190  0.01786156 9.857493e-01
#> gene4     gene4 4.401789e+00    1.398448752 1.0751193  1.30073826 1.933481e-01
#> gene6     gene6 5.339168e+00   -0.329714559 1.1076218 -0.29767792 7.659490e-01
#> gene11   gene11 5.303104e+00   -0.881232651 1.0341587 -0.85212516 3.941446e-01
#> gene14   gene14 9.317441e-01   -2.674101675 2.4887082 -1.07449386 2.826014e-01
#> gene19   gene19 5.334582e+00   -1.442426872 1.0779596 -1.33810845 1.808611e-01
#> gene20   gene20 4.317545e+00   -1.317286953 1.1907288 -1.10628633 2.686026e-01
#> gene24   gene24 5.868818e+00    0.214031520 0.9651463  0.22176069 8.245002e-01
#> gene33   gene33 1.022026e+00   -1.924924820 2.1565181 -0.89260777 3.720673e-01
#> gene38   gene38 2.993797e+00   -3.670503059 1.6273602 -2.25549509 2.410227e-02
#> gene41   gene41 3.432829e+00    2.016215292 2.4984333  0.80699183 4.196712e-01
#> gene46   gene46 9.993230e-01   -1.918826836 2.2978995 -0.83503514 4.036979e-01
#> gene48   gene48 4.831450e-01   -0.372072237 3.1934611 -0.11651065 9.072478e-01
#> gene53   gene53 3.922485e+00    0.925507409 1.1379508  0.81331059 4.160400e-01
#> gene54   gene54 9.995528e-02    0.378454923 3.5338115  0.10709539 9.147133e-01
#> gene57   gene57 5.631026e+00   -0.463135206 0.9737950 -0.47559826 6.343606e-01
#> gene59   gene59 1.051824e-01    0.378454923 3.5338115  0.10709539 9.147133e-01
#> gene60   gene60 3.531173e+00   -1.418628024 1.3227928 -1.07244915 2.835183e-01
#> gene65   gene65 1.024091e-01    0.378454923 3.5338115  0.10709539 9.147133e-01
#> gene71   gene71 2.348889e+00   -2.697097620 1.7667151 -1.52661715 1.268562e-01
#> gene72   gene72 4.153361e-01   -2.393963962 3.4827989 -0.68736783 4.918510e-01
#> gene84   gene84 2.434433e+00    0.249333090 1.6433952  0.15171828 8.794091e-01
#> gene91   gene91 1.051824e-01    0.378454923 3.5338115  0.10709539 9.147133e-01
#> gene92   gene92 2.314845e+00    0.141932170 1.4014394  0.10127599 9.193314e-01
#> gene98   gene98 2.248318e+00   -1.352776659 1.7692801 -0.76459156 4.445148e-01
#>                 padj
#> gene12  2.560881e-18
#> gene87  2.578696e-07
#> gene89  2.906709e-07
#> gene83  6.710716e-07
#> gene77  3.377273e-05
#> gene82  5.236269e-05
#> gene39  3.938410e-04
#> gene30  2.269946e-03
#> gene70  2.269946e-03
#> gene29  4.071367e-03
#> gene97  6.296603e-03
#> gene88  6.640009e-03
#> gene2   7.245936e-03
#> gene79  3.160262e-02
#> gene61  3.555335e-02
#> gene100 4.182813e-02
#> gene45  4.769720e-02
#> gene58  4.769720e-02
#> gene75  4.769720e-02
#> gene81  5.893590e-02
#> gene27  6.376948e-02
#> gene40  6.999621e-02
#> gene15  8.137384e-02
#> gene32  8.137384e-02
#> gene90  8.137384e-02
#> gene8   9.885234e-02
#> gene42  9.885234e-02
#> gene74  1.060012e-01
#> gene69  1.446701e-01
#> gene28  2.410640e-01
#> gene55  2.410640e-01
#> gene96  2.410640e-01
#> gene63  2.426849e-01
#> gene26  2.447282e-01
#> gene66  2.530955e-01
#> gene93  2.530955e-01
#> gene94  2.530955e-01
#> gene5   2.617955e-01
#> gene22  2.617955e-01
#> gene9   2.849684e-01
#> gene35  2.849684e-01
#> gene36  2.849684e-01
#> gene44  2.849684e-01
#> gene18  2.914979e-01
#> gene17  3.081479e-01
#> gene68  3.092344e-01
#> gene13  3.280725e-01
#> gene31  3.591010e-01
#> gene95  3.674687e-01
#> gene50  4.141846e-01
#> gene47  4.387788e-01
#> gene56  4.554095e-01
#> gene62  4.554095e-01
#> gene10  4.816340e-01
#> gene85  5.566704e-01
#> gene99  5.566704e-01
#> gene80  5.657917e-01
#> gene64  5.965589e-01
#> gene16  5.966578e-01
#> gene21  5.966578e-01
#> gene51  5.966578e-01
#> gene86  5.966578e-01
#> gene43  6.307279e-01
#> gene3   6.362948e-01
#> gene34  6.784080e-01
#> gene67  6.784080e-01
#> gene73  6.876111e-01
#> gene7   7.393963e-01
#> gene23  7.430859e-01
#> gene78  7.480082e-01
#> gene49  7.522568e-01
#> gene52  7.522568e-01
#> gene37  8.667793e-01
#> gene25  8.877910e-01
#> gene76  9.917994e-01
#> gene1             NA
#> gene4             NA
#> gene6             NA
#> gene11            NA
#> gene14            NA
#> gene19            NA
#> gene20            NA
#> gene24            NA
#> gene33            NA
#> gene38            NA
#> gene41            NA
#> gene46            NA
#> gene48            NA
#> gene53            NA
#> gene54            NA
#> gene57            NA
#> gene59            NA
#> gene60            NA
#> gene65            NA
#> gene71            NA
#> gene72            NA
#> gene84            NA
#> gene91            NA
#> gene92            NA
#> gene98            NA