Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult

A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n = 100, m = 8, betaSD = 1)
dds <- DESeq2::DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- DESeq2::results(dds)
deseqresult2tbl(res)
#>              id    baseMean log2FoldChange     lfcSE        stat       pvalue
#> gene77   gene77  70.4341980    2.985608057 0.4347724  6.86706016 6.553844e-12
#> gene66   gene66 237.2894564    2.229108103 0.3872859  5.75571704 8.627479e-09
#> gene96   gene96  16.8584970   -3.581574770 0.7328249 -4.88735390 1.022003e-06
#> gene78   gene78  18.5506455   -2.514736773 0.6488481 -3.87569415 1.063212e-04
#> gene16   gene16  26.4730335   -1.908185697 0.5370983 -3.55276781 3.812008e-04
#> gene9     gene9 470.4577435    1.112322443 0.3311575  3.35889273 7.825544e-04
#> gene14   gene14  50.0270775    1.496858084 0.4379011  3.41825597 6.302381e-04
#> gene80   gene80  26.1798770   -1.666417715 0.4987022 -3.34150876 8.332439e-04
#> gene7     gene7  46.0559455    1.557376215 0.4812742  3.23594401 1.212412e-03
#> gene25   gene25  46.4512459   -1.727947564 0.5428293 -3.18322457 1.456446e-03
#> gene34   gene34  93.7767157    1.329830992 0.4445275  2.99156040 2.775556e-03
#> gene62   gene62  34.7568476   -1.376238113 0.4570636 -3.01104324 2.603518e-03
#> gene99   gene99 104.6317639    1.210966012 0.4064440  2.97941680 2.887976e-03
#> gene24   gene24  19.6473642    1.796911319 0.6695578  2.68372835 7.280624e-03
#> gene81   gene81 151.2431029   -0.883136155 0.3303994 -2.67293493 7.519083e-03
#> gene54   gene54  11.4204917   -1.798883517 0.7051691 -2.55099607 1.074155e-02
#> gene69   gene69  24.5322682   -1.404931474 0.5482947 -2.56236581 1.039618e-02
#> gene91   gene91  69.7830530    0.922027718 0.3626260  2.54264114 1.100182e-02
#> gene55   gene55 227.3155312   -0.832387461 0.3311232 -2.51383015 1.194279e-02
#> gene40   gene40   8.4650943    2.242403217 0.9216369  2.43306583 1.497158e-02
#> gene17   gene17  42.4815724   -0.975032487 0.4092437 -2.38252291 1.719446e-02
#> gene21   gene21  10.5419576    2.105032259 0.9214890  2.28438138 2.234913e-02
#> gene57   gene57 190.7201964    0.679680265 0.2949123  2.30468606 2.118416e-02
#> gene94   gene94  21.4873340   -1.383186065 0.6075324 -2.27672824 2.280246e-02
#> gene72   gene72  15.0522525    1.742206420 0.7805956  2.23189369 2.562199e-02
#> gene90   gene90  13.9837613   -1.358138844 0.6134795 -2.21382932 2.684052e-02
#> gene63   gene63   4.2579437   -2.956529418 1.3450148 -2.19813900 2.793920e-02
#> gene64   gene64   4.7385357   -2.278106716 1.0788946 -2.11151922 3.472771e-02
#> gene49   gene49  19.4629933   -1.310218253 0.6446340 -2.03249952 4.210311e-02
#> gene93   gene93  95.5174571    0.760835922 0.3767139  2.01966487 4.341816e-02
#> gene23   gene23  27.8722018    0.799264765 0.4346319  1.83894639 6.592307e-02
#> gene68   gene68  67.4177899    0.806726804 0.4673176  1.72629225 8.429485e-02
#> gene76   gene76   5.7032113    1.864480901 1.0764130  1.73212403 8.325148e-02
#> gene4     gene4   7.4967566   -1.264240124 0.7733405 -1.63477807 1.020956e-01
#> gene37   gene37   5.5413624   -1.603903331 0.9830321 -1.63158795 1.027663e-01
#> gene6     gene6  57.5278658    0.681580080 0.4271862  1.59551053 1.105981e-01
#> gene33   gene33 134.5907464    0.608812847 0.3937866  1.54604771 1.220930e-01
#> gene28   gene28  16.3070648   -0.981861588 0.6651918 -1.47605777 1.399284e-01
#> gene30   gene30 155.4546007   -0.497679928 0.3363776 -1.47952734 1.389994e-01
#> gene83   gene83   7.2319769    1.454609121 0.9856381  1.47580443 1.399964e-01
#> gene87   gene87  77.1307722    0.472669564 0.3226966  1.46474927 1.429893e-01
#> gene8     gene8  12.7729369    1.151411173 0.7934064  1.45122493 1.467172e-01
#> gene5     gene5  38.5509141   -0.653629426 0.4811631 -1.35843630 1.743253e-01
#> gene43   gene43   4.0050931    1.882649221 1.4031304  1.34174931 1.796773e-01
#> gene61   gene61  54.1062718   -0.636601507 0.4703413 -1.35348837 1.758997e-01
#> gene92   gene92   1.2078611    2.883039453 2.2420205  1.28591130 1.984740e-01
#> gene15   gene15  53.3329751   -0.561693462 0.4535253 -1.23850530 2.155288e-01
#> gene36   gene36  54.0364466   -0.463804749 0.3885613 -1.19364629 2.326164e-01
#> gene13   gene13   3.1920108    1.684599378 1.4612234  1.15286912 2.489641e-01
#> gene71   gene71   3.0015859   -1.599566269 1.3766268 -1.16194617 2.452573e-01
#> gene31   gene31 102.3020555   -0.416117603 0.3698334 -1.12514891 2.605259e-01
#> gene27   gene27   4.1005219   -1.392757038 1.2659098 -1.10020247 2.712439e-01
#> gene67   gene67  58.3471058   -0.450668165 0.4095013 -1.10052922 2.711016e-01
#> gene35   gene35  18.4679991   -0.605146161 0.6056649 -0.99914359 3.177251e-01
#> gene44   gene44   9.4048909    0.739837839 0.7483487  0.98862710 3.228456e-01
#> gene97   gene97  36.1826965   -0.470618554 0.4753961 -0.98995045 3.221983e-01
#> gene1     gene1  13.6634529   -0.643593542 0.7287370 -0.88316299 3.771482e-01
#> gene39   gene39  34.1017983    0.406544594 0.4619812  0.88000251 3.788580e-01
#> gene45   gene45  13.6680474   -0.696347858 0.7635177 -0.91202586 3.617551e-01
#> gene50   gene50  15.9595766    0.541098792 0.6169633  0.87703561 3.804673e-01
#> gene95   gene95  10.2124155    0.759433258 0.8595095  0.88356585 3.769306e-01
#> gene38   gene38   2.5211327    1.154589638 1.3399565  0.86166204 3.888735e-01
#> gene59   gene59   1.6924411    1.345483065 1.6583355  0.81134552 4.171673e-01
#> gene22   gene22  23.5556204    0.418208266 0.5441890  0.76849825 4.421912e-01
#> gene74   gene74  41.5335649    0.418987550 0.5458947  0.76752449 4.427697e-01
#> gene18   gene18  27.5260478   -0.349298153 0.5071814 -0.68870456 4.910092e-01
#> gene75   gene75   2.4897822    0.892297336 1.3155123  0.67828885 4.975886e-01
#> gene2     gene2  15.0492638    0.376640744 0.5861309  0.64258808 5.204914e-01
#> gene52   gene52   9.4047808   -0.508531008 0.8407738 -0.60483688 5.452874e-01
#> gene73   gene73   3.0479597   -0.677965591 1.1987029 -0.56558265 5.716775e-01
#> gene79   gene79  28.8458487   -0.255406116 0.4810243 -0.53096301 5.954444e-01
#> gene32   gene32  27.8575731    0.254921081 0.6171772  0.41304358 6.795747e-01
#> gene70   gene70  17.3211934    0.287803954 0.7213568  0.39897588 6.899110e-01
#> gene26   gene26   3.9894087    0.345046799 1.0725770  0.32169885 7.476809e-01
#> gene29   gene29   3.5906792    0.396514069 1.1343677  0.34954633 7.266792e-01
#> gene51   gene51  31.6872155    0.152449319 0.4616565  0.33022243 7.412319e-01
#> gene65   gene65  12.8287979   -0.292977794 0.8798932 -0.33296970 7.391572e-01
#> gene3     gene3   2.7135166    0.373442208 1.5043056  0.24824890 8.039418e-01
#> gene10   gene10   7.4197610    0.145114690 0.9354963  0.15512053 8.767263e-01
#> gene11   gene11  32.8754805   -0.066253496 0.4906597 -0.13502941 8.925886e-01
#> gene19   gene19   9.4097766    0.064206824 0.7453456  0.08614370 9.313522e-01
#> gene41   gene41  11.4651387   -0.048369875 0.6834264 -0.07077555 9.435764e-01
#> gene47   gene47 147.7457350    0.049348425 0.4424156  0.11154313 9.111857e-01
#> gene48   gene48  16.0946752    0.093990927 0.6646806  0.14140767 8.875479e-01
#> gene53   gene53  16.9776542   -0.091818216 0.6146879 -0.14937372 8.812587e-01
#> gene56   gene56  76.3463919   -0.037050804 0.4116096 -0.09001442 9.282758e-01
#> gene82   gene82   1.9098112   -0.211999985 1.5648007 -0.13548050 8.922320e-01
#> gene84   gene84  13.1345750    0.136153382 0.7306686  0.18634081 8.521775e-01
#> gene86   gene86  40.2159386    0.044263785 0.4410650  0.10035660 9.200612e-01
#> gene88   gene88   4.9183226    0.058136559 0.9694800  0.05996674 9.521821e-01
#> gene89   gene89  21.2091717    0.053562724 0.5936610  0.09022443 9.281089e-01
#> gene98   gene98   5.2562935   -0.070771201 1.0407589 -0.06799961 9.457859e-01
#> gene100 gene100   7.8246965   -0.081584599 0.7511863 -0.10860767 9.135137e-01
#> gene20   gene20 147.9869617    0.009242355 0.3566820  0.02591203 9.793275e-01
#> gene85   gene85  43.8502584   -0.015803743 0.5224570 -0.03024889 9.758686e-01
#> gene12   gene12   0.7041239    1.004422754 2.5063388  0.40075299 6.886020e-01
#> gene42   gene42   0.5217984    2.576972843 3.1950548  0.80655044 4.199255e-01
#> gene46   gene46   0.8846101   -1.468638698 2.5399218 -0.57822202 5.631142e-01
#> gene58   gene58   0.3588256   -0.712277786 3.4298188 -0.20767213 8.354850e-01
#> gene60   gene60   0.0000000             NA        NA          NA           NA
#>                 padj
#> gene77  6.226151e-10
#> gene66  4.098053e-07
#> gene96  3.236342e-05
#> gene78  2.525128e-03
#> gene16  7.242815e-03
#> gene9   9.894771e-03
#> gene14  9.894771e-03
#> gene80  9.894771e-03
#> gene7   1.279768e-02
#> gene25  1.383624e-02
#> gene34  2.110444e-02
#> gene62  2.110444e-02
#> gene99  2.110444e-02
#> gene24  4.762086e-02
#> gene81  4.762086e-02
#> gene54  5.806514e-02
#> gene69  5.806514e-02
#> gene91  5.806514e-02
#> gene55  5.971397e-02
#> gene40  7.111500e-02
#> gene17  7.778447e-02
#> gene21  9.025973e-02
#> gene57  9.025973e-02
#> gene94  9.025973e-02
#> gene72  9.736356e-02
#> gene90  9.807114e-02
#> gene63  9.830460e-02
#> gene64  1.178262e-01
#> gene49  1.374908e-01
#> gene93  1.374908e-01
#> gene23  2.020223e-01
#> gene68  2.426670e-01
#> gene76  2.426670e-01
#> gene4   2.789371e-01
#> gene37  2.789371e-01
#> gene6   2.918561e-01
#> gene33  3.134821e-01
#> gene28  3.313168e-01
#> gene30  3.313168e-01
#> gene83  3.313168e-01
#> gene87  3.313168e-01
#> gene8   3.318604e-01
#> gene5   3.793187e-01
#> gene43  3.793187e-01
#> gene61  3.793187e-01
#> gene92  4.098920e-01
#> gene15  4.356432e-01
#> gene36  4.603865e-01
#> gene13  4.730318e-01
#> gene71  4.730318e-01
#> gene31  4.852934e-01
#> gene27  4.861919e-01
#> gene67  4.861919e-01
#> gene35  5.476845e-01
#> gene44  5.476845e-01
#> gene97  5.476845e-01
#> gene1   5.925310e-01
#> gene39  5.925310e-01
#> gene45  5.925310e-01
#> gene50  5.925310e-01
#> gene95  5.925310e-01
#> gene38  5.958546e-01
#> gene59  6.290618e-01
#> gene22  6.471250e-01
#> gene74  6.471250e-01
#> gene18  7.055360e-01
#> gene75  7.055360e-01
#> gene2   7.271571e-01
#> gene52  7.507580e-01
#> gene73  7.758481e-01
#> gene79  7.967214e-01
#> gene32  8.966610e-01
#> gene70  8.978294e-01
#> gene26  9.224634e-01
#> gene29  9.224634e-01
#> gene51  9.224634e-01
#> gene65  9.224634e-01
#> gene3   9.726592e-01
#> gene10  9.726592e-01
#> gene11  9.726592e-01
#> gene19  9.726592e-01
#> gene41  9.726592e-01
#> gene47  9.726592e-01
#> gene48  9.726592e-01
#> gene53  9.726592e-01
#> gene56  9.726592e-01
#> gene82  9.726592e-01
#> gene84  9.726592e-01
#> gene86  9.726592e-01
#> gene88  9.726592e-01
#> gene89  9.726592e-01
#> gene98  9.726592e-01
#> gene100 9.726592e-01
#> gene20  9.793275e-01
#> gene85  9.793275e-01
#> gene12            NA
#> gene42            NA
#> gene46            NA
#> gene58            NA
#> gene60            NA