Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult
A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n=100, m=8, betaSD = 1)
dds <- DESeq2::DESeq(dds)

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> final dispersion estimates

#> fitting model and testing

res <- DESeq2::results(dds)
deseqresult2tbl(res)

#>              id     baseMean log2FoldChange     lfcSE        stat       pvalue
#> gene18   gene18  147.7672115     1.81180737 0.3686343  4.91491833 8.881948e-07
#> gene60   gene60   46.2117002    -2.30736434 0.4869602 -4.73830204 2.155164e-06
#> gene49   gene49  158.0878830    -1.77872348 0.4018392 -4.42645570 9.579405e-06
#> gene53   gene53   56.4539309     2.14780789 0.4837266  4.44012781 8.990547e-06
#> gene4     gene4   11.7942965     4.55988006 1.1262978  4.04855647 5.153450e-05
#> gene37   gene37  176.3726912    -1.67093529 0.4226210 -3.95374438 7.693760e-05
#> gene92   gene92  175.9733980     1.54740357 0.4007370  3.86139464 1.127416e-04
#> gene36   gene36   92.8407969     1.93239884 0.5048087  3.82798263 1.291979e-04
#> gene48   gene48   30.4723121     2.34723328 0.6916218  3.39381066 6.892733e-04
#> gene46   gene46    8.5793311    -2.77782874 0.9074062 -3.06128469 2.203894e-03
#> gene21   gene21   24.8595855     1.56326481 0.5231209  2.98834303 2.804945e-03
#> gene15   gene15    9.5788500     2.21770592 0.8080480  2.74452250 6.059900e-03
#> gene8     gene8   16.5528205    -1.67399665 0.6457407 -2.59236652 9.531816e-03
#> gene54   gene54    6.0058185    -2.51750607 1.0180315 -2.47291581 1.340157e-02
#> gene30   gene30   43.4305173     1.34138186 0.5800370  2.31257990 2.074575e-02
#> gene42   gene42   55.0885987     1.08580918 0.4770517  2.27608278 2.284106e-02
#> gene43   gene43   12.9858958    -1.49771300 0.6545149 -2.28827934 2.212126e-02
#> gene75   gene75  116.1426810     1.22606727 0.5432433  2.25693937 2.401186e-02
#> gene79   gene79   47.6292374    -1.09029621 0.4769335 -2.28605506 2.225104e-02
#> gene85   gene85   33.0411647    -1.02506194 0.4367920 -2.34679668 1.893558e-02
#> gene81   gene81   24.0312805    -1.48459644 0.6713788 -2.21126493 2.701750e-02
#> gene29   gene29   63.0771671    -1.07928627 0.5376468 -2.00742607 4.470432e-02
#> gene40   gene40    9.5288356     1.88458618 0.9792124  1.92459391 5.428018e-02
#> gene66   gene66   16.3941665    -1.15943246 0.6083937 -1.90572715 5.668562e-02
#> gene26   gene26   15.8285239    -1.38367754 0.7668934 -1.80426328 7.119004e-02
#> gene16   gene16   32.1330846    -0.76434535 0.4382109 -1.74424098 8.111708e-02
#> gene71   gene71   12.7817906     1.36388808 0.7851506  1.73710370 8.236887e-02
#> gene5     gene5    6.0124241    -1.81829129 1.1130781 -1.63357032 1.023491e-01
#> gene13   gene13    6.4063855    -1.49460009 0.9147094 -1.63396170 1.022669e-01
#> gene23   gene23   13.4648587     1.21480571 0.7114228  1.70757212 8.771577e-02
#> gene31   gene31   24.5704309     1.25738995 0.7636200  1.64661746 9.963670e-02
#> gene62   gene62    5.1692910    -2.12569227 1.2953884 -1.64096904 1.008038e-01
#> gene95   gene95   21.4640620     0.87067439 0.5231249  1.66437202 9.603811e-02
#> gene86   gene86  221.9549512     0.55484371 0.3529658  1.57194767 1.159627e-01
#> gene35   gene35   26.4328448    -0.68474822 0.4983452 -1.37404403 1.694280e-01
#> gene65   gene65   32.4237187    -0.75311719 0.5684238 -1.32492197 1.851970e-01
#> gene96   gene96   30.6283666     0.76853965 0.5997525  1.28142795 2.000434e-01
#> gene3     gene3    5.2306502    -1.11346675 0.9780366 -1.13847140 2.549237e-01
#> gene11   gene11    6.4005978     1.06668978 0.9803392  1.08808233 2.765588e-01
#> gene17   gene17 1391.8628616     0.37605638 0.3382549  1.11175445 2.662437e-01
#> gene70   gene70    6.6501426     0.92780707 0.8680806  1.06880293 2.851585e-01
#> gene89   gene89   36.6558955    -0.57596358 0.5403255 -1.06595672 2.864432e-01
#> gene67   gene67   27.9234651    -0.53718367 0.5176427 -1.03774985 2.993865e-01
#> gene12   gene12   90.2425723    -0.43366393 0.4371815 -0.99195399 3.212200e-01
#> gene47   gene47   20.3033965     0.74597537 0.7629593  0.97773943 3.282032e-01
#> gene51   gene51   56.4652546     0.44687893 0.4635993  0.96393350 3.350793e-01
#> gene56   gene56   94.4751277     0.37403462 0.3851316  0.97118661 3.314554e-01
#> gene69   gene69   59.7861354     0.44439149 0.4564299  0.97362491 3.302428e-01
#> gene100 gene100   19.9579969     0.68554977 0.7297941  0.93937427 3.475386e-01
#> gene80   gene80   31.0274383    -0.50935446 0.5708139 -0.89233023 3.722160e-01
#> gene39   gene39   11.3378252    -0.58412561 0.7063621 -0.82694923 4.082658e-01
#> gene7     gene7   21.7636045     0.43516949 0.5566033  0.78183063 4.343141e-01
#> gene45   gene45   14.7497426    -0.58419746 0.7227839 -0.80826018 4.189408e-01
#> gene61   gene61   20.1452134    -0.46556041 0.5944918 -0.78312333 4.335547e-01
#> gene63   gene63   14.9836946    -0.50678268 0.6713637 -0.75485561 4.503356e-01
#> gene33   gene33   19.5336530    -0.41848060 0.5806479 -0.72071319 4.710860e-01
#> gene88   gene88   52.1231411    -0.33328987 0.4546401 -0.73308512 4.635065e-01
#> gene41   gene41    6.3492340    -0.61221774 0.9174292 -0.66731876 5.045685e-01
#> gene59   gene59   13.1651336     0.64283316 0.9564520  0.67210186 5.015189e-01
#> gene87   gene87    4.4440153     0.67929922 0.9713311  0.69934881 4.843341e-01
#> gene99   gene99  275.6150416    -0.29462152 0.4295066 -0.68595335 4.927425e-01
#> gene1     gene1    3.8392185     0.67384885 1.2030111  0.56013519 5.753872e-01
#> gene10   gene10    6.5887951     0.62179064 1.0837341  0.57374832 5.661381e-01
#> gene57   gene57    6.8034194    -0.48962702 0.8833023 -0.55431422 5.793638e-01
#> gene93   gene93   18.8449121     0.34888736 0.6464980  0.53965731 5.894334e-01
#> gene27   gene27    4.9642290    -0.65003837 1.3018191 -0.49933079 6.175464e-01
#> gene78   gene78    9.6375469     0.45507594 0.8673976  0.52464514 5.998299e-01
#> gene90   gene90   22.9211061     0.32104338 0.6281354  0.51110540 6.092773e-01
#> gene77   gene77    5.4123345     0.43505735 0.9837035  0.44226474 6.582976e-01
#> gene58   gene58   68.3003757     0.16705903 0.4459982  0.37457331 7.079778e-01
#> gene52   gene52   11.3296541     0.31599323 0.8812153  0.35858801 7.199033e-01
#> gene44   gene44   63.0369191     0.13387207 0.4195890  0.31905522 7.496846e-01
#> gene94   gene94   86.0570634     0.12124901 0.3763524  0.32216880 7.473248e-01
#> gene25   gene25   31.6298448    -0.13956872 0.5122069 -0.27248506 7.852491e-01
#> gene55   gene55   12.8749623    -0.18290707 0.6649339 -0.27507558 7.832582e-01
#> gene76   gene76   59.9514264    -0.14486450 0.4937379 -0.29340363 7.692136e-01
#> gene72   gene72   32.7347702     0.13428254 0.5482607  0.24492460 8.065148e-01
#> gene9     gene9    7.2478328    -0.16791632 0.8979562 -0.18699833 8.516619e-01
#> gene20   gene20  127.6055875    -0.07480987 0.3854973 -0.19406071 8.461283e-01
#> gene22   gene22   26.8980437    -0.10288668 0.5614556 -0.18324987 8.546020e-01
#> gene32   gene32   41.7556954    -0.04366768 0.4501030 -0.09701707 9.227128e-01
#> gene83   gene83    7.4533000     0.02898913 0.9773727  0.02966026 9.763380e-01
#> gene2     gene2    1.4282806    -2.41236030 1.8007965 -1.33960741 1.803730e-01
#> gene6     gene6    0.0000000             NA        NA          NA           NA
#> gene14   gene14    2.6542864     0.29120929 1.4432203  0.20177743 8.400907e-01
#> gene19   gene19    1.0868084     3.60628939 2.6374248  1.36735250 1.715148e-01
#> gene24   gene24    1.6695261     0.34604530 1.6676830  0.20750065 8.356189e-01
#> gene28   gene28    1.8363302    -1.97980220 1.9197535 -1.03127936 3.024098e-01
#> gene34   gene34    1.2689141    -0.32804349 1.9576886 -0.16756674 8.669241e-01
#> gene38   gene38    0.2383993    -1.48142894 3.5322017 -0.41940666 6.749190e-01
#> gene50   gene50    0.2383993    -1.48142894 3.5322017 -0.41940666 6.749190e-01
#> gene64   gene64    3.6161705    -1.73328483 1.6008438 -1.08273203 2.789274e-01
#> gene68   gene68    1.0560216    -3.61265286 2.9826735 -1.21121299 2.258138e-01
#> gene73   gene73    5.8940837    -1.12781617 1.3617290 -0.82822367           NA
#> gene74   gene74    2.2734338    -2.44088593 1.8927070 -1.28962697 1.971802e-01
#> gene82   gene82    1.9571366     1.15281811 1.6087710  0.71658311 4.736314e-01
#> gene84   gene84    0.8252266    -2.34184359 2.2461682 -1.04259495 2.971359e-01
#> gene91   gene91    3.0676579    -2.13899624 1.3472277 -1.58770210 1.123537e-01
#> gene97   gene97    2.1419029    -1.41655534 1.4225627 -0.99577706 3.193585e-01
#> gene98   gene98    3.5736711     0.86849726 1.4523690  0.59798663 5.498489e-01
#>                 padj
#> gene18  7.283197e-05
#> gene60  8.836173e-05
#> gene49  1.963778e-04
#> gene53  1.963778e-04
#> gene4   8.451658e-04
#> gene37  1.051481e-03
#> gene92  1.320688e-03
#> gene36  1.324278e-03
#> gene48  6.280046e-03
#> gene46  1.807193e-02
#> gene21  2.090959e-02
#> gene15  4.140932e-02
#> gene8   6.012376e-02
#> gene54  7.849492e-02
#> gene30  9.844864e-02
#> gene42  9.844864e-02
#> gene43  9.844864e-02
#> gene75  9.844864e-02
#> gene79  9.844864e-02
#> gene85  9.844864e-02
#> gene81  1.054969e-01
#> gene29  1.666252e-01
#> gene40  1.935206e-01
#> gene66  1.936759e-01
#> gene26  2.335033e-01
#> gene16  2.501573e-01
#> gene71  2.501573e-01
#> gene5   2.543220e-01
#> gene13  2.543220e-01
#> gene23  2.543220e-01
#> gene31  2.543220e-01
#> gene62  2.543220e-01
#> gene95  2.543220e-01
#> gene86  2.796747e-01
#> gene35  3.969456e-01
#> gene65  4.218377e-01
#> gene96  4.433394e-01
#> gene3   5.500985e-01
#> gene11  5.592463e-01
#> gene17  5.592463e-01
#> gene70  5.592463e-01
#> gene89  5.592463e-01
#> gene67  5.709232e-01
#> gene12  5.724271e-01
#> gene47  5.724271e-01
#> gene51  5.724271e-01
#> gene56  5.724271e-01
#> gene69  5.724271e-01
#> gene100 5.815952e-01
#> gene80  6.104342e-01
#> gene39  6.564274e-01
#> gene7   6.595140e-01
#> gene45  6.595140e-01
#> gene61  6.595140e-01
#> gene63  6.714095e-01
#> gene33  6.777027e-01
#> gene88  6.777027e-01
#> gene41  6.782725e-01
#> gene59  6.782725e-01
#> gene87  6.782725e-01
#> gene99  6.782725e-01
#> gene1   7.423099e-01
#> gene10  7.423099e-01
#> gene57  7.423099e-01
#> gene93  7.435929e-01
#> gene27  7.446883e-01
#> gene78  7.446883e-01
#> gene90  7.446883e-01
#> gene77  7.823247e-01
#> gene58  8.293455e-01
#> gene52  8.314376e-01
#> gene44  8.421115e-01
#> gene94  8.421115e-01
#> gene25  8.472424e-01
#> gene55  8.472424e-01
#> gene76  8.472424e-01
#> gene72  8.588859e-01
#> gene9   8.759670e-01
#> gene20  8.759670e-01
#> gene22  8.759670e-01
#> gene32  9.341044e-01
#> gene83  9.763380e-01
#> gene2             NA
#> gene6             NA
#> gene14            NA
#> gene19            NA
#> gene24            NA
#> gene28            NA
#> gene34            NA
#> gene38            NA
#> gene50            NA
#> gene64            NA
#> gene68            NA
#> gene73            NA
#> gene74            NA
#> gene82            NA
#> gene84            NA
#> gene91            NA
#> gene97            NA
#> gene98            NA