Generate a tidy table with the results of DESeq

Generate a tidy table with the results of DESeq

deseqresult2tbl(deseqresult)

Arguments

deseqresult

A DESeqResults object

Value

A "tidy" data.frame with all genes

Examples

# with simulated data...
library(DESeq2)
dds <- DESeq2::makeExampleDESeqDataSet(n = 100, m = 8, betaSD = 1)
dds <- DESeq2::DESeq(dds)
#> estimating size factors
#> estimating dispersions
#> gene-wise dispersion estimates
#> mean-dispersion relationship
#> final dispersion estimates
#> fitting model and testing
res <- DESeq2::results(dds)
deseqresult2tbl(res)
#>              id    baseMean log2FoldChange     lfcSE          stat       pvalue
#> gene47   gene47  64.7762804    2.286081774 0.4167258  5.4858175870 4.115614e-08
#> gene3     gene3 155.7304572    1.665922140 0.3712774  4.4870010059 7.223273e-06
#> gene1     gene1  31.3270214    2.015243208 0.5082715  3.9648951964 7.342820e-05
#> gene4     gene4  49.7264546    1.736466317 0.4399869  3.9466317561 7.925831e-05
#> gene44   gene44  13.6250466    3.976189614 0.9905479  4.0141317820 5.966501e-05
#> gene53   gene53 144.1872737    1.484410640 0.3778940  3.9281139486 8.561465e-05
#> gene78   gene78  51.5327343    2.153844428 0.5334035  4.0379271400 5.392560e-05
#> gene10   gene10  14.9763070   -2.483430472 0.6590029 -3.7684664631 1.642535e-04
#> gene24   gene24  32.4571859    2.278611513 0.6147939  3.7063012646 2.103081e-04
#> gene83   gene83  16.0785061    2.208892793 0.6559403  3.3675213072 7.584717e-04
#> gene19   gene19  60.5175214    1.334800700 0.4252538  3.1388328687 1.696222e-03
#> gene35   gene35  28.0707489   -1.554404603 0.5280446 -2.9436992756 3.243150e-03
#> gene18   gene18  26.6905200   -1.431004083 0.4905785 -2.9169724114 3.534470e-03
#> gene21   gene21  36.6692418    1.074868723 0.4445765  2.4177360273 1.561740e-02
#> gene36   gene36  13.9224710   -1.568100947 0.7072192 -2.2172769901 2.660417e-02
#> gene38   gene38   8.9567943   -1.888489138 0.8449133 -2.2351276959 2.540897e-02
#> gene82   gene82  42.2702617    1.161944632 0.5200355  2.2343562595 2.545964e-02
#> gene95   gene95  25.7431144   -1.094423677 0.4934219 -2.2180280554 2.655292e-02
#> gene30   gene30  57.2806875    0.982731218 0.4595482  2.1384725152 3.247841e-02
#> gene7     gene7   9.6514558   -1.604720701 0.7638785 -2.1007537307 3.566259e-02
#> gene11   gene11  31.3962269    0.973133184 0.4856373  2.0038272612 4.508857e-02
#> gene37   gene37 205.0379947   -0.815097894 0.4093515 -1.9911929267 4.645968e-02
#> gene88   gene88   9.0926485   -1.579519501 0.7808672 -2.0227759833 4.309625e-02
#> gene6     gene6  58.2307183   -0.790889093 0.4108924 -1.9248082985 5.425335e-02
#> gene65   gene65  15.2437112   -1.248781312 0.6490381 -1.9240493948 5.434839e-02
#> gene32   gene32  24.2904654    1.182262362 0.6312445  1.8729073285 6.108118e-02
#> gene87   gene87  90.4547031    0.811557100 0.4380556  1.8526351174 6.393468e-02
#> gene23   gene23  72.3840578    0.698679765 0.3908167  1.7877427160 7.381753e-02
#> gene81   gene81   8.7219151    1.553424373 0.9039311  1.7185207170 8.570168e-02
#> gene12   gene12  10.3156586   -1.414588583 0.8511545 -1.6619644308 9.651991e-02
#> gene96   gene96  21.0446648   -1.008643330 0.6064714 -1.6631341673 9.628558e-02
#> gene25   gene25  13.4377494    1.184288440 0.8755699  1.3525915637 1.761861e-01
#> gene15   gene15  16.8880550   -0.967579161 0.7355097 -1.3155219063 1.883346e-01
#> gene17   gene17  92.1360367    0.436291654 0.3535465  1.2340433340 2.171868e-01
#> gene67   gene67   9.7319613   -0.888000930 0.7347646 -1.2085516053 2.268352e-01
#> gene73   gene73  49.9283078    0.570085543 0.4707010  1.2111416698 2.258411e-01
#> gene94   gene94 164.5664680   -0.488072778 0.3905436 -1.2497268482 2.113993e-01
#> gene69   gene69  26.3296353    0.565629653 0.4849710  1.1663163457 2.434866e-01
#> gene2     gene2  11.6804832   -0.721484401 0.6450493 -1.1184949875 2.633556e-01
#> gene27   gene27  21.4091558    1.442453991 1.2851653  1.1223878900 2.616976e-01
#> gene77   gene77   8.7812390   -0.850400794 0.7641427 -1.1128822447 2.657590e-01
#> gene34   gene34  36.3564927    0.536418381 0.5065344  1.0589969277 2.896012e-01
#> gene97   gene97  15.0156516   -0.765524712 0.7228247 -1.0590738699 2.895661e-01
#> gene54   gene54  22.2949696    0.711825316 0.7169823  0.9928073595 3.208038e-01
#> gene22   gene22  38.5527069    0.411957710 0.4350472  0.9469265065 3.436762e-01
#> gene26   gene26  12.0005029   -0.632343916 0.7141194 -0.8854876987 3.758937e-01
#> gene72   gene72   9.1407779    0.656268407 0.7299088  0.8991101046 3.685940e-01
#> gene89   gene89  40.7047147   -0.377223738 0.4241414 -0.8893820217 3.737978e-01
#> gene90   gene90   9.2933416    0.805290131 0.8715883  0.9239340408 3.555207e-01
#> gene49   gene49   8.2055431    0.672346501 0.8301742  0.8098860806 4.180057e-01
#> gene57   gene57  13.4414707   -0.551770808 0.7205018 -0.7658145845 4.437866e-01
#> gene60   gene60  22.7253355   -0.352384409 0.4802675 -0.7337252679 4.631162e-01
#> gene79   gene79  65.1385407    0.235910140 0.4030879  0.5852572888 5.583747e-01
#> gene68   gene68  58.8514058    0.261203642 0.4675662  0.5586452862 5.764038e-01
#> gene84   gene84  14.0344078   -0.340043121 0.6649749 -0.5113623677 6.090973e-01
#> gene76   gene76  15.5733065    0.287519938 0.6330186  0.4542045748 6.496816e-01
#> gene52   gene52  21.4655674   -0.242866327 0.5700735 -0.4260263469 6.700886e-01
#> gene16   gene16  50.1046220   -0.116928477 0.3758988 -0.3110637183 7.557522e-01
#> gene33   gene33  14.1744602    0.217875947 0.6522311  0.3340471503 7.383440e-01
#> gene46   gene46  26.3624924    0.182664956 0.5127250  0.3562629904 7.216436e-01
#> gene58   gene58  34.4866856   -0.178605444 0.4904225 -0.3641868578 7.157185e-01
#> gene86   gene86   9.9789698   -0.227956336 0.7097704 -0.3211691383 7.480822e-01
#> gene100 gene100  25.5015872    0.148689544 0.6046933  0.2458924756 8.057655e-01
#> gene28   gene28  16.9353530   -0.092954306 0.7424904 -0.1251926008 9.003711e-01
#> gene13   gene13   8.3976074   -0.073581889 0.7941639 -0.0926532822 9.261790e-01
#> gene9     gene9  14.8193109    0.048171756 0.6701001  0.0718873983 9.426915e-01
#> gene5     gene5   8.1744772    0.416632496 0.8563778  0.4865054685 6.266088e-01
#> gene8     gene8   0.8256541   -1.201589299 2.4453782 -0.4913715620 6.231637e-01
#> gene14   gene14   1.4486189   -0.346827498 1.6249853 -0.2134342423 8.309883e-01
#> gene20   gene20   5.3783711    2.526621921 1.0400337  2.4293655478 1.512527e-02
#> gene29   gene29   0.0000000             NA        NA            NA           NA
#> gene31   gene31   5.4006972    1.394993431 1.1487982  1.2143067503 2.246306e-01
#> gene39   gene39   8.1482467    1.640659045 0.9072921  1.8083030197 7.055935e-02
#> gene40   gene40   2.4936161    0.740820656 1.4437009  0.5131399840 6.078534e-01
#> gene41   gene41   4.7848432   -1.116468726 1.0720603 -1.0414234639 2.976791e-01
#> gene42   gene42   6.3939221   -1.187294891 1.0673267 -1.1124006577 2.659659e-01
#> gene43   gene43   2.1359654   -1.698611354 1.4645737 -1.1597991691 2.461306e-01
#> gene45   gene45   2.0021292   -1.122482268 1.8541816 -0.6053788309 5.449273e-01
#> gene48   gene48   1.7752955    0.765225019 1.7593107  0.4349573019 6.635934e-01
#> gene50   gene50   4.9783618   -2.468332740 1.0536892 -2.3425624719 1.915183e-02
#> gene51   gene51   2.7154473   -3.286252546 1.5442729 -2.1280258439 3.333494e-02
#> gene55   gene55   1.0606032   -1.679159817 2.2472068 -0.7472208832 4.549302e-01
#> gene56   gene56   0.2504688    1.585952808 3.5318978  0.4490370051 6.534050e-01
#> gene59   gene59   2.6823554   -1.662559175 1.4581144 -1.1402117226 2.541981e-01
#> gene61   gene61   0.6167578    1.846330756 3.1406778  0.5878765311 5.566152e-01
#> gene62   gene62   2.3077445    0.572226468 1.3048019  0.4385543022 6.609845e-01
#> gene63   gene63   4.3723576    1.691022764 1.0557947  1.6016587230 1.092311e-01
#> gene64   gene64   4.9033950   -0.532942685 1.0143506 -0.5254028508 5.993032e-01
#> gene66   gene66   7.0207216   -0.955419559 0.9327128 -1.0243448416 3.056724e-01
#> gene70   gene70   4.6151970    1.615271006 1.0489976  1.5398233054 1.236034e-01
#> gene71   gene71   4.2675955   -0.678773700 1.1276185 -0.6019533001 5.472052e-01
#> gene74   gene74   1.3013197   -2.921733673 1.9767603 -1.4780414361 1.393967e-01
#> gene75   gene75   0.9448660   -0.546507639 2.1580758 -0.2532383914 8.000840e-01
#> gene80   gene80   3.7785022    0.397759935 1.2913131  0.3080274866 7.580614e-01
#> gene85   gene85   5.7339990    0.072190988 1.0056883  0.0717826673 9.427749e-01
#> gene91   gene91   2.9863657   -0.001478425 3.2776002 -0.0004510693 9.996401e-01
#> gene92   gene92   1.6574864    0.572356735 1.9247649  0.2973644846 7.661883e-01
#> gene93   gene93   5.6136892   -1.618536356 1.1771492 -1.3749628342 1.691430e-01
#> gene98   gene98   5.4101880   -1.907245077 1.0642302 -1.7921358136 7.311121e-02
#> gene99   gene99   7.0921067   -0.309248349 0.8132626 -0.3802564383 7.037551e-01
#>                 padj
#> gene47  2.716305e-06
#> gene3   2.383680e-04
#> gene1   8.072238e-04
#> gene4   8.072238e-04
#> gene44  8.072238e-04
#> gene53  8.072238e-04
#> gene78  8.072238e-04
#> gene10  1.355092e-03
#> gene24  1.542260e-03
#> gene83  5.005913e-03
#> gene19  1.017733e-02
#> gene35  1.783732e-02
#> gene18  1.794423e-02
#> gene21  7.362489e-02
#> gene36  9.754862e-02
#> gene38  9.754862e-02
#> gene82  9.754862e-02
#> gene95  9.754862e-02
#> gene30  1.128197e-01
#> gene7   1.176865e-01
#> gene11  1.333191e-01
#> gene37  1.333191e-01
#> gene88  1.333191e-01
#> gene6   1.434798e-01
#> gene65  1.434798e-01
#> gene32  1.550522e-01
#> gene87  1.562848e-01
#> gene23  1.739985e-01
#> gene81  1.950452e-01
#> gene12  2.054940e-01
#> gene96  2.054940e-01
#> gene25  3.633839e-01
#> gene15  3.766691e-01
#> gene17  4.046249e-01
#> gene67  4.046249e-01
#> gene73  4.046249e-01
#> gene94  4.046249e-01
#> gene69  4.228977e-01
#> gene2   4.278072e-01
#> gene27  4.278072e-01
#> gene77  4.278072e-01
#> gene34  4.445041e-01
#> gene97  4.445041e-01
#> gene54  4.812058e-01
#> gene22  5.040585e-01
#> gene26  5.063057e-01
#> gene72  5.063057e-01
#> gene89  5.063057e-01
#> gene90  5.063057e-01
#> gene49  5.517675e-01
#> gene57  5.743121e-01
#> gene60  5.878013e-01
#> gene79  6.953346e-01
#> gene68  7.044936e-01
#> gene84  7.309168e-01
#> gene76  7.656962e-01
#> gene52  7.758921e-01
#> gene16  8.045104e-01
#> gene33  8.045104e-01
#> gene46  8.045104e-01
#> gene58  8.045104e-01
#> gene86  8.045104e-01
#> gene100 8.441353e-01
#> gene28  9.285077e-01
#> gene13  9.404279e-01
#> gene9   9.426915e-01
#> gene5             NA
#> gene8             NA
#> gene14            NA
#> gene20            NA
#> gene29            NA
#> gene31            NA
#> gene39            NA
#> gene40            NA
#> gene41            NA
#> gene42            NA
#> gene43            NA
#> gene45            NA
#> gene48            NA
#> gene50            NA
#> gene51            NA
#> gene55            NA
#> gene56            NA
#> gene59            NA
#> gene61            NA
#> gene62            NA
#> gene63            NA
#> gene64            NA
#> gene66            NA
#> gene70            NA
#> gene71            NA
#> gene74            NA
#> gene75            NA
#> gene80            NA
#> gene85            NA
#> gene91            NA
#> gene92            NA
#> gene93            NA
#> gene98            NA
#> gene99            NA